==> ClimateIndices/Src/06_BayesianLinearHierarchicalModels.R <==
library(ncdf4)
library(chron)
library(tidyverse)
library(kohonen) # fitting
library(aweSOM) # plotting
library(SOMbrero) # plotting
library(paletteer) #colors
library(PNWColors) #more colors
library(here) #navigating folders
library(sf)
library(rnaturalearth)
library(rnaturalearthdata)
library(maps)       #basic mapping functions and some data
library(rstan)
library(mapdata)    #some additional hires data
library(maptools)   #useful tools such as reading shapefiles
library(mapproj)
library(PBSmapping)
library(bayestestR)
library(dplyr)
set.seed(1234)

##### Writing Bayes DFA model function ####
warmups <- 1000
total_iterations <- 4000
max_treedepth <-  10
n_chains <-  3
n_cores <- 3
adapt_delta <- 0.95
col<-pnw_palette("Sunset2",3,type="discrete")

bayeslinmod <- function(dat, ind,xyz, period) {
  N <- length(period)
  NP <- as.numeric(length(unique(period)))
  P<- as.numeric(as.factor(period))
  K <- 1
  y <-xyz
  x<-dat[ind]
  data <- list(N = N, #total number of opservations
               NP=NP, #total number of time periods
               P = P, #time period pointer vector
               K = K,#number of covariates, starting with one but can add for final model structure
               x=x, 
               y=y #response variable
  )
  
  bhfit <- stan(
    file = here::here("Src/BayesianLinearHierarchicalModels.stan"),
    data = data,
    chains = n_chains,
    warmup = warmups,
    iter = total_iterations,
    cores = n_cores,
    refresh = 250,
    control = list(max_treedepth = max_treedepth,
                   adapt_delta = adapt_delta))
  posterior<-data.frame(summary(bhfit, prob=c(0.025, 0.25,0.75, 0.975, 0.1, 0.9))$summary)
  scaled<-NULL
  alphaP<-posterior[1:NP,]%>%
    add_column(
      period=rep(seq(1,NP),1))
  
  betaP<-posterior[seq(NP+1,NP*2),]%>%
    add_column(
      period=rep(seq(1,NP),1))
  n<- total_iterations
  for(i in 1:NP){
    tempalpha <- rnorm(n, alphaP$mean[i],alphaP$sd[i])
    tempbeta <- rnorm(n, betaP$mean[i],betaP$sd[i])
    scaled <-rbind(scaled,cbind(tempalpha, tempbeta))
  }
  scaled.anom<<- scaled%>%
    bind_cols(period=rep(rep(seq(1,NP), each = n),1), Index=rep(colnames(dat[ind]), n*NP),
              yfirst=rep(min(dat$Year_lag), n*NP),ylast=rep(max(dat$Year_lag), n*NP))%>%
    rename(alpha=tempalpha, beta = tempbeta)
}

postplot <- function(dat, param){
  ggplot(dat, aes(x = param, fill = as.factor(period), group=as.factor(period))) +
    theme_bw() +
    facet_wrap(.~Index, ncol = 4, scales='free') +
    geom_density(alpha = 0.7) +
    scale_fill_manual(values = c(col[1], col[2], col[3])) +
    #theme(legend.title = element_blank(), legend.position = 'top', legend.key.size = unit(3, 'mm')) +
    geom_vline(xintercept = 0, lty = 2) +
    labs(x = "Slope",
         y = "Posterior density")
}

##### DFA Trend Model Runs ####
climate_dat_cop<-readRDS(here('data/physical/climate_dat_cop.rds'))
dfa<-readRDS(here('data/physical/climate_dat_dfa.rds'))%>%
  mutate(period=ifelse(Year_lag<1989,1,ifelse(Year_lag>2012,3,2)))
dat <- dfa%>%dplyr::select(estimate, Year_lag, lower, upper,trend,season)%>%
  mutate(Year_lag = Year_lag+1)%>%
  rename(estimateoffset1=estimate, loweroffset1=lower, upperoffset1=upper)%>%
  left_join(dfa, by=c('trend', 'Year_lag','season'))

climpdo <- climate_dat_cop%>%filter(season=="Spring")%>%
  dplyr::select(Year_lag, region, seasonal_PDO,  seasonal_NPGO,  seasonal_ONI,  seasonal_NPH)%>%
  distinct()%>%
  filter(Year_lag<2024)
data <- climate_dat_cop%>%filter(season=="Summer")%>%
  dplyr::select(Year_lag, region, seasonal_copepod_northern, 
         seasonal_copepod_southern)%>%
  left_join(climpdo)%>%
  distinct()%>%
  mutate(period=ifelse(Year_lag<2013,2,3))%>%
  filter(Year_lag>1996)%>%
  dplyr::select(Year_lag, seasonal_copepod_northern,seasonal_copepod_southern,
         seasonal_NPH,seasonal_NPGO,seasonal_PDO,seasonal_ONI)%>%
  rename(NPH=seasonal_NPH,NPGO=seasonal_NPGO,PDO=seasonal_PDO,ONI=seasonal_ONI)%>%
  pivot_longer(!c(Year_lag,  NPH, NPGO,PDO,ONI), 
               names_to = "trend", values_to = "estimate")%>%
  mutate(season="Spring")%>%
  pivot_longer(!c(Year_lag, season, estimate,trend), 
               names_to = "Index_Name", values_to = "Index_Value")



dat.long<- dat%>%filter(season=="Spring")%>%
  dplyr::select(Year_lag, season, estimate,trend,estimateoffset1,
         seasonal_NPH,seasonal_NPGO,seasonal_PDO,seasonal_ONI)%>%
  rename(NPH=seasonal_NPH,NPGO=seasonal_NPGO,PDO=seasonal_PDO,ONI=seasonal_ONI)%>%
  distinct()%>%
  pivot_longer(!c(Year_lag, season, estimate,trend), 
               names_to = "Index_Name", values_to = "Index_Value")

dat.lm<-data%>%
  bind_rows(dat.long)%>%
  mutate(period=ifelse(Year_lag<1989,1,ifelse(Year_lag>2012,3,2)))%>%
  filter(trend!="SEA")%>%
  mutate(trend=ifelse(trend=="CALCOFI","CALCOFI (SCC)",
                      ifelse(trend=="RREAS", "RREAS (CCC)", 
                          ifelse(trend=="seasonal_copepod_southern","Southern Copepod (NCC)","Northern Copepod (NCC)"))))%>%
  mutate(trend = fct_relevel(trend, "Northern Copepod (NCC)", 
                              "Southern Copepod (NCC)","RREAS (CCC)","CALCOFI (SCC)"))

survey.lm<-ggplot(data = dat.lm%>%filter(period!=4), aes(y = estimate, x =Index_Value,col=as.factor(period))) +
  facet_grid(Index_Name~trend, scales='free') +
   geom_point(aes(col=as.factor(period))) +
  # geom_text(aes(label=Year_lag,col=as.factor(period))) +
  geom_smooth(method = "lm", se = FALSE, aes(col=as.factor(period))) +
 # geom_smooth(method = "lm", se = FALSE, col='grey') +
  scale_y_continuous(name = "Index of Abundance") +
  scale_color_manual(values =  col[1:3], name="Period",labels=c('1967 - 1988', '1989 - 2012','2013 - 2023'))+
  theme_bw()+
  xlab("Climate Index Value")+
  theme(plot.title = element_text(hjust = 0.5))+
  ggtitle("Spring")
survey.lm



pdf(file = "Output/Figures/biological.lm.pdf",   # The directory you want to save the file in
    width = 8.5, # The width of the plot in inches
    height = 6)
survey.lm
dev.off()
#### CALCOFI ####
######Spring #####

data <- dat%>%filter(season=="Spring", trend=="CALCOFI")%>%
  distinct()%>%
  filter(Year_lag<2023)%>%
  add_column(full=1)



CALCOFI <-NULL
columns<-c(which(colnames(data) == "seasonal_PDO"), which(colnames(data) == "seasonal_NPGO"),
           which(colnames(data) == "seasonal_NPH"),which(colnames(data) == "seasonal_ONI"))
for(i in 1:length(columns)){
  bayeslinmod(data, columns[i],  data$estimate, data$period)
  CALCOFI <-rbind(CALCOFI,scaled.anom)
}

CALCOFI<-CALCOFI%>%mutate(survey="CALCOFI",
Index=ifelse(Index=="seasonal_PDO", "PDO", ifelse(Index=="seasonal_NPGO", "NPGO",
                                                  ifelse(Index=="seasonal_NPH", "NPH","ONI"))))

postplot(CALCOFI, CALCOFI$beta)
postplot(CALCOFI, CALCOFI$alpha)

CALCOFI_full <-NULL
columns<-c(which(colnames(data) == "seasonal_PDO"), which(colnames(data) == "seasonal_NPGO"),
           which(colnames(data) == "seasonal_NPH"),which(colnames(data) == "seasonal_ONI"))
for(i in 1:length(columns)){
  bayeslinmod(data, columns[i],  data$estimate, data$full)
  CALCOFI_full <-rbind(CALCOFI_full,scaled.anom)
}

CALCOFI_full<-CALCOFI_full%>%mutate(survey="CALCOFI",
Index=ifelse(Index=="seasonal_PDO", "PDO", ifelse(Index=="seasonal_NPGO", "NPGO",
                                                  ifelse(Index=="seasonal_NPH", "NPH","ONI"))))%>%
  mutate(period=4)

ggplot(CALCOFI_full, aes(x = beta),fill = as.factor(period), group=as.factor(period)) +
    theme_bw() +
    facet_wrap(.~Index, ncol = 4, scales='free') +
    geom_density(alpha = 0.7) +
    scale_fill_manual(values = c(col[1])) +
    #theme(legend.title = element_blank(), legend.position = 'top', legend.key.size = unit(3, 'mm')) +
    geom_vline(xintercept = 0, lty = 2) +
    labs(x = "Slope",
         y = "Posterior density")


CALCOFI<- bind_rows(CALCOFI,CALCOFI_full)%>%
  mutate(region="SCC",Season="Spring", lag = 0)%>%
  filter(period!=4)

CALCOFI_beta<-ggplot(CALCOFI, aes(x = beta,fill = as.factor(period), group=as.factor(period))) +
    theme_bw() +
    facet_wrap(.~Index, ncol = 4, scales='free') +
    geom_density(alpha = 0.7) +
    scale_fill_manual(values = c(col[1],col[2],col[3],'grey')) +
    #theme(legend.title = element_blank(), legend.position = 'top', legend.key.size = unit(3, 'mm')) +
    geom_vline(xintercept = 0, lty = 2) +
    labs(x = "Slope",
         y = "Posterior density")

index.names <- unique(CALCOFI$Index)
period.names <- unique(CALCOFI$period)

overlap.CALCOFI <- NA
  for(i in 1:4){
    temp <- CALCOFI%>%filter(Index==index.names[i])%>%dplyr::select(beta, period)
    ov1 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(beta),temp%>%filter(period==2)%>%dplyr::select(beta)), period1=c(1), period2=c(2), Index=index.names[i])
    ov2 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(beta),temp%>%filter(period==3)%>%dplyr::select(beta)), period1=c(1), period2=c(3), Index=index.names[i])
    ov3 <- data.frame(ov=overlap(temp%>%filter(period==3)%>%dplyr::select(beta),temp%>%filter(period==2)%>%dplyr::select(beta)), period1=c(3), period2=c(2), Index=index.names[i])
 #   ov4 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(beta),temp%>%filter(period==3)%>%dplyr::select(beta)), period1=c(4), period2=c(3), Index=index.names[i])
 ##   ov5 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(beta),temp%>%filter(period==1)%>%dplyr::select(beta)), period1=c(4), period2=c(2), Index=index.names[i])
  #  ov6 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(beta),temp%>%filter(period==2)%>%dplyr::select(beta)), period1=c(4), period2=c(1), Index=index.names[i])

    temp2<-rbind(ov1,ov2,ov3)
    overlap.CALCOFI <-rbind(temp2,overlap.CALCOFI)
  }
overlap.CALCOFI.beta<-overlap.CALCOFI%>%mutate(Survey="CALCOFI", region="SCC", season="Spring", offset=0) 

overlap.CALCOFI <- NA
for(i in 1:4){
  temp <- CALCOFI%>%filter(Index==index.names[i])%>%dplyr::select(alpha, period)
  ov1 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(alpha),temp%>%filter(period==2)%>%dplyr::select(alpha)), period1=c(1), period2=c(2), Index=index.names[i])
  ov2 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(alpha),temp%>%filter(period==3)%>%dplyr::select(alpha)), period1=c(1), period2=c(3), Index=index.names[i])
  ov3 <- data.frame(ov=overlap(temp%>%filter(period==3)%>%dplyr::select(alpha),temp%>%filter(period==2)%>%dplyr::select(alpha)), period1=c(3), period2=c(2), Index=index.names[i])
  #   ov4 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(alpha),temp%>%filter(period==3)%>%dplyr::select(alpha)), period1=c(4), period2=c(3), Index=index.names[i])
  ##   ov5 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(alpha),temp%>%filter(period==1)%>%dplyr::select(alpha)), period1=c(4), period2=c(2), Index=index.names[i])
  #  ov6 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(alpha),temp%>%filter(period==2)%>%dplyr::select(alpha)), period1=c(4), period2=c(1), Index=index.names[i])
  
  temp2<-rbind(ov1,ov2,ov3)
  overlap.CALCOFI <-rbind(temp2,overlap.CALCOFI)
}
overlap.CALCOFI.alpha<-overlap.CALCOFI%>%mutate(Survey="CALCOFI", region="SCC", season="Spring", offset=0) 

#running models with 1-year offset

CALCOFIoffset <-NULL
columns<-c(which(colnames(data) == "seasonal_PDO"), which(colnames(data) == "seasonal_NPGO"),
           which(colnames(data) == "seasonal_NPH"),which(colnames(data) == "seasonal_ONI"))
for(i in 1:length(columns)){
  bayeslinmod(data, columns[i],  data$estimateoffset1, data$period)
  CALCOFIoffset <-rbind(CALCOFIoffset,scaled.anom)
}
CALCOFIoffset<-CALCOFIoffset%>%mutate(survey="CALCOFI",
                          Index=ifelse(Index=="seasonal_PDO", "PDO", ifelse(Index=="seasonal_NPGO", "NPGO",
                                                                            ifelse(Index=="seasonal_NPH", "NPH","ONI"))))
  
postplot(CALCOFIoffset, CALCOFIoffset$beta)
postplot(CALCOFIoffset, CALCOFIoffset$alpha)

CALCOFIoffset_full <-NULL
columns<-c(which(colnames(data) == "seasonal_PDO"), which(colnames(data) == "seasonal_NPGO"),
           which(colnames(data) == "seasonal_NPH"),which(colnames(data) == "seasonal_ONI"))
for(i in 1:length(columns)){
  bayeslinmod(data, columns[i],  data$estimate, data$full)
  CALCOFIoffset_full <-rbind(CALCOFIoffset_full,scaled.anom)
}

CALCOFIoffset_full<-CALCOFIoffset_full%>%mutate(survey="CALCOFI",
Index=ifelse(Index=="seasonal_PDO", "PDO", ifelse(Index=="seasonal_NPGO", "NPGO",
                                                  ifelse(Index=="seasonal_NPH", "NPH","ONI"))))%>%
  mutate(period=4)

ggplot(CALCOFIoffset_full, aes(x = beta),fill = as.factor(period), group=as.factor(period)) +
    theme_bw() +
    facet_wrap(.~Index, ncol = 4, scales='free') +
    geom_density(alpha = 0.7) +
    scale_fill_manual(values = c(col[1])) +
    #theme(legend.title = element_blank(), legend.position = 'top', legend.key.size = unit(3, 'mm')) +
    geom_vline(xintercept = 0, lty = 2) +
    labs(x = "Slope",
         y = "Posterior density")


CALCOFIoffset<- bind_rows(CALCOFIoffset,CALCOFIoffset_full)%>%
  mutate(region="SCC",Season="Spring", lag = 1)

index.names <- unique(CALCOFIoffset$Index)
period.names <- unique(CALCOFIoffset$period)
overlap.CALCOFIoffset <- NA
  for(i in 1:4){
    temp <- CALCOFIoffset%>%filter(Index==index.names[i])%>%dplyr::select(beta, period)
    ov1 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(beta),temp%>%filter(period==2)%>%dplyr::select(beta)), period1=c(1), period2=c(2), Index=index.names[i])
    ov2 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(beta),temp%>%filter(period==3)%>%dplyr::select(beta)), period1=c(1), period2=c(3), Index=index.names[i])
    ov3 <- data.frame(ov=overlap(temp%>%filter(period==3)%>%dplyr::select(beta),temp%>%filter(period==2)%>%dplyr::select(beta)), period1=c(3), period2=c(2), Index=index.names[i])
    ov4 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(beta),temp%>%filter(period==3)%>%dplyr::select(beta)), period1=c(4), period2=c(3), Index=index.names[i])
    ov5 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(beta),temp%>%filter(period==1)%>%dplyr::select(beta)), period1=c(4), period2=c(2), Index=index.names[i])
    ov6 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(beta),temp%>%filter(period==2)%>%dplyr::select(beta)), period1=c(4), period2=c(1), Index=index.names[i])

    temp2<-rbind(ov1,ov2,ov3, ov4, ov5, ov6)
    overlap.CALCOFIoffset <-rbind(temp2,overlap.CALCOFIoffset)
  }
overlap.CALCOFIoffset.beta<-overlap.CALCOFIoffset%>%mutate(Survey="CALCOFI", region="SCC", season="Spring", offset=1) 
overlap.CALCOFIoffset <- NA
for(i in 1:4){
  temp <- CALCOFIoffset%>%filter(Index==index.names[i])%>%dplyr::select(alpha, period)
  ov1 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(alpha),temp%>%filter(period==2)%>%dplyr::select(alpha)), period1=c(1), period2=c(2), Index=index.names[i])
  ov2 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(alpha),temp%>%filter(period==3)%>%dplyr::select(alpha)), period1=c(1), period2=c(3), Index=index.names[i])
  ov3 <- data.frame(ov=overlap(temp%>%filter(period==3)%>%dplyr::select(alpha),temp%>%filter(period==2)%>%dplyr::select(alpha)), period1=c(3), period2=c(2), Index=index.names[i])
  ov4 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(alpha),temp%>%filter(period==3)%>%dplyr::select(alpha)), period1=c(4), period2=c(3), Index=index.names[i])
  ov5 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(alpha),temp%>%filter(period==1)%>%dplyr::select(alpha)), period1=c(4), period2=c(2), Index=index.names[i])
  ov6 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(alpha),temp%>%filter(period==2)%>%dplyr::select(alpha)), period1=c(4), period2=c(1), Index=index.names[i])
  
  temp2<-rbind(ov1,ov2,ov3, ov4, ov5, ov6)
  overlap.CALCOFIoffset <-rbind(temp2,overlap.CALCOFIoffset)
}
overlap.CALCOFIoffset.alpha<-overlap.CALCOFIoffset%>%mutate(Survey="CALCOFI", region="SCC", season="Spring", offset=1) 

###### Winter #####

data <- dat%>%filter(season=="Winter", trend=="CALCOFI")%>%
  distinct()%>%
  filter(Year_lag<2023)%>%
  add_column(full=1)

CALCOFI_W <-NULL
columns<-c(which(colnames(data) == "seasonal_PDO"), which(colnames(data) == "seasonal_NPGO"),
           which(colnames(data) == "seasonal_NPH"),which(colnames(data) == "seasonal_ONI"))
for(i in 1:length(columns)){
  bayeslinmod(data, columns[i],  data$estimate, data$period)
  CALCOFI_W <-rbind(CALCOFI_W,scaled.anom)
}

CALCOFI_W<-CALCOFI_W%>%mutate(survey="CALCOFI",
Index=ifelse(Index=="seasonal_PDO", "PDO", ifelse(Index=="seasonal_NPGO", "NPGO",
                                                  ifelse(Index=="seasonal_NPH", "NPH","ONI"))))

postplot(CALCOFI_W, CALCOFI_W$beta)
postplot(CALCOFI_W, CALCOFI_W$alpha)

CALCOFI_W_full <-NULL
columns<-c(which(colnames(data) == "seasonal_PDO"), which(colnames(data) == "seasonal_NPGO"),
           which(colnames(data) == "seasonal_NPH"),which(colnames(data) == "seasonal_ONI"))
for(i in 1:length(columns)){
  bayeslinmod(data, columns[i],  data$estimate, data$full)
  CALCOFI_W_full <-rbind(CALCOFI_W_full,scaled.anom)
}

CALCOFI_W_full<-CALCOFI_W_full%>%mutate(survey="CALCOFI",
Index=ifelse(Index=="seasonal_PDO", "PDO", ifelse(Index=="seasonal_NPGO", "NPGO",
                                                  ifelse(Index=="seasonal_NPH", "NPH","ONI"))))%>%
  mutate(period=4)

ggplot(CALCOFI_W_full, aes(x = beta),fill = as.factor(period), group=as.factor(period)) +
    theme_bw() +
    facet_wrap(.~Index, ncol = 4, scales='free') +
    geom_density(alpha = 0.7) +
    scale_fill_manual(values = c(col[1])) +
    #theme(legend.title = element_blank(), legend.position = 'top', legend.key.size = unit(3, 'mm')) +
    geom_vline(xintercept = 0, lty = 2) +
    labs(x = "Slope",
         y = "Posterior density")


CALCOFI_W<- bind_rows(CALCOFI_W,CALCOFI_W_full)%>%
  mutate(region="SCC",Season="Winter", lag = 0)

index.names <- unique(CALCOFI_W$Index)
period.names <- unique(CALCOFI_W$period)
overlap.CALCOFI_W <- NA
  for(i in 1:4){
    temp <- CALCOFI_W%>%filter(Index==index.names[i])%>%dplyr::select(beta, period)
    ov1 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(beta),temp%>%filter(period==2)%>%dplyr::select(beta)), period1=c(1), period2=c(2), Index=index.names[i])
    ov2 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(beta),temp%>%filter(period==3)%>%dplyr::select(beta)), period1=c(1), period2=c(3), Index=index.names[i])
    ov3 <- data.frame(ov=overlap(temp%>%filter(period==3)%>%dplyr::select(beta),temp%>%filter(period==2)%>%dplyr::select(beta)), period1=c(3), period2=c(2), Index=index.names[i])
    ov4 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(beta),temp%>%filter(period==3)%>%dplyr::select(beta)), period1=c(4), period2=c(3), Index=index.names[i])
    ov5 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(beta),temp%>%filter(period==1)%>%dplyr::select(beta)), period1=c(4), period2=c(2), Index=index.names[i])
    ov6 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(beta),temp%>%filter(period==2)%>%dplyr::select(beta)), period1=c(4), period2=c(1), Index=index.names[i])

    temp2<-rbind(ov1,ov2,ov3, ov4, ov5, ov6)
    overlap.CALCOFI_W <-rbind(temp2,overlap.CALCOFI_W)
  }
overlap.CALCOFI_W.beta<-overlap.CALCOFI_W%>%mutate(Survey="CALCOFI", region="SCC", season="Winter", offset=0) 

overlap.CALCOFI_W <- NA
for(i in 1:4){
  temp <- CALCOFI_W%>%filter(Index==index.names[i])%>%dplyr::select(alpha, period)
  ov1 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(alpha),temp%>%filter(period==2)%>%dplyr::select(alpha)), period1=c(1), period2=c(2), Index=index.names[i])
  ov2 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(alpha),temp%>%filter(period==3)%>%dplyr::select(alpha)), period1=c(1), period2=c(3), Index=index.names[i])
  ov3 <- data.frame(ov=overlap(temp%>%filter(period==3)%>%dplyr::select(alpha),temp%>%filter(period==2)%>%dplyr::select(alpha)), period1=c(3), period2=c(2), Index=index.names[i])
  ov4 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(alpha),temp%>%filter(period==3)%>%dplyr::select(alpha)), period1=c(4), period2=c(3), Index=index.names[i])
  ov5 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(alpha),temp%>%filter(period==1)%>%dplyr::select(alpha)), period1=c(4), period2=c(2), Index=index.names[i])
  ov6 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(alpha),temp%>%filter(period==2)%>%dplyr::select(alpha)), period1=c(4), period2=c(1), Index=index.names[i])
  
  temp2<-rbind(ov1,ov2,ov3, ov4, ov5, ov6)
  overlap.CALCOFI_W <-rbind(temp2,overlap.CALCOFI_W)
}
overlap.CALCOFI_W.alpha<-overlap.CALCOFI_W%>%mutate(Survey="CALCOFI", region="SCC", season="Winter", offset=0) 


#running models with 1-year offset

CALCOFI_Woffset <-NULL
columns<-c(which(colnames(data) == "seasonal_PDO"), which(colnames(data) == "seasonal_NPGO"),
           which(colnames(data) == "seasonal_NPH"),which(colnames(data) == "seasonal_ONI"))
for(i in 1:length(columns)){
  bayeslinmod(data, columns[i],  data$estimateoffset1, data$period)
  CALCOFI_Woffset <-rbind(CALCOFI_Woffset,scaled.anom)
}
CALCOFI_Woffset<-CALCOFI_Woffset%>%mutate(survey="CALCOFI",
                          Index=ifelse(Index=="seasonal_PDO", "PDO", ifelse(Index=="seasonal_NPGO", "NPGO",
                                                                            ifelse(Index=="seasonal_NPH", "NPH","ONI"))))
  
postplot(CALCOFI_Woffset, CALCOFI_Woffset$beta)
postplot(CALCOFI_Woffset, CALCOFI_Woffset$alpha)

CALCOFI_Woffset_full <-NULL
columns<-c(which(colnames(data) == "seasonal_PDO"), which(colnames(data) == "seasonal_NPGO"),
           which(colnames(data) == "seasonal_NPH"),which(colnames(data) == "seasonal_ONI"))
for(i in 1:length(columns)){
  bayeslinmod(data, columns[i],  data$estimate, data$full)
  CALCOFI_Woffset_full <-rbind(CALCOFI_Woffset_full,scaled.anom)
}

CALCOFI_Woffset_full<-CALCOFI_Woffset_full%>%mutate(survey="CALCOFI",
Index=ifelse(Index=="seasonal_PDO", "PDO", ifelse(Index=="seasonal_NPGO", "NPGO",
                                                  ifelse(Index=="seasonal_NPH", "NPH","ONI"))))%>%
  mutate(period=4)

ggplot(CALCOFI_Woffset_full, aes(x = beta),fill = as.factor(period), group=as.factor(period)) +
    theme_bw() +
    facet_wrap(.~Index, ncol = 4, scales='free') +
    geom_density(alpha = 0.7) +
    scale_fill_manual(values = c(col[1])) +
    #theme(legend.title = element_blank(), legend.position = 'top', legend.key.size = unit(3, 'mm')) +
    geom_vline(xintercept = 0, lty = 2) +
    labs(x = "Slope",
         y = "Posterior density")


CALCOFI_Woffset<- bind_rows(CALCOFI_Woffset,CALCOFI_Woffset_full)%>%
  mutate(region="SCC",Season="Winter", lag = 1)

index.names <- unique(CALCOFI_Woffset$Index)
period.names <- unique(CALCOFI_Woffset$period)
overlap.CALCOFI_Woffset <- NA
 for(i in 1:4){
    temp <- CALCOFI_Woffset%>%filter(Index==index.names[i])%>%dplyr::select(beta, period)
    ov1 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(beta),temp%>%filter(period==2)%>%dplyr::select(beta)), period1=c(1), period2=c(2), Index=index.names[i])
    ov2 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(beta),temp%>%filter(period==3)%>%dplyr::select(beta)), period1=c(1), period2=c(3), Index=index.names[i])
    ov3 <- data.frame(ov=overlap(temp%>%filter(period==3)%>%dplyr::select(beta),temp%>%filter(period==2)%>%dplyr::select(beta)), period1=c(3), period2=c(2), Index=index.names[i])
   # ov4 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(beta),temp%>%filter(period==3)%>%dplyr::select(beta)), period1=c(4), period2=c(3), Index=index.names[i])
  #  ov5 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(beta),temp%>%filter(period==1)%>%dplyr::select(beta)), period1=c(4), period2=c(2), Index=index.names[i])
  #  ov6 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(beta),temp%>%filter(period==2)%>%dplyr::select(beta)), period1=c(4), period2=c(1), Index=index.names[i])

    temp2<-rbind(ov1,ov2,ov3, ov4, ov5, ov6)
    overlap.CALCOFI_Woffset <-rbind(temp2,overlap.CALCOFI_Woffset)
  }
overlap.CALCOFI_Woffset.beta<-overlap.CALCOFI_Woffset%>%mutate(Survey="CALCOFI", region="SCC", season="Winter", offset=1)



#### RREAS ####
######Spring #####

data <- dat%>%filter(season=="Spring", trend=="RREAS")%>%
  distinct()%>%
  filter(Year_lag<2023)%>%
  add_column(full=1)

RREAS <-NULL
columns<-c(which(colnames(data) == "seasonal_PDO"), which(colnames(data) == "seasonal_NPGO"),
           which(colnames(data) == "seasonal_NPH"),which(colnames(data) == "seasonal_ONI"))
for(i in 1:length(columns)){
  bayeslinmod(data, columns[i],  data$estimate, data$period)
  RREAS <-rbind(RREAS,scaled.anom)
}

RREAS<-RREAS%>%mutate(survey="RREAS",
Index=ifelse(Index=="seasonal_PDO", "PDO", ifelse(Index=="seasonal_NPGO", "NPGO",
                                                  ifelse(Index=="seasonal_NPH", "NPH","ONI"))))

postplot(RREAS, RREAS$beta)
postplot(RREAS, RREAS$alpha)

RREAS_full <-NULL
columns<-c(which(colnames(data) == "seasonal_PDO"), which(colnames(data) == "seasonal_NPGO"),
           which(colnames(data) == "seasonal_NPH"),which(colnames(data) == "seasonal_ONI"))
for(i in 1:length(columns)){
  bayeslinmod(data, columns[i],  data$estimate, data$full)
  RREAS_full <-rbind(RREAS_full,scaled.anom)
}

RREAS_full<-RREAS_full%>%mutate(survey="RREAS",
Index=ifelse(Index=="seasonal_PDO", "PDO", ifelse(Index=="seasonal_NPGO", "NPGO",
                                                  ifelse(Index=="seasonal_NPH", "NPH","ONI"))))%>%
  mutate(period=4)

ggplot(RREAS, aes(x = beta),fill = as.factor(period), group=as.factor(period)) +
    theme_bw() +
    facet_wrap(.~Index, ncol = 4, scales='free') +
    geom_density(alpha = 0.7) +
    scale_fill_manual(values = c(col[1])) +
    #theme(legend.title = element_blank(), legend.position = 'top', legend.key.size = unit(3, 'mm')) +
    geom_vline(xintercept = 0, lty = 2) +
    labs(x = "Slope",
         y = "Posterior density")


RREAS<- bind_rows(RREAS,RREAS_full)%>%
  mutate(region="CCC",Season="Spring", lag = 0)

index.names <- unique(RREAS$Index)
period.names <- unique(RREAS$period)
overlap.RREAS <- NA
for(i in 1:4){
    temp <- RREAS%>%filter(Index==index.names[i])%>%dplyr::select(beta, period)
    ov1 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(beta),temp%>%filter(period==2)%>%dplyr::select(beta)), period1=c(1), period2=c(2), Index=index.names[i])
    ov2 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(beta),temp%>%filter(period==3)%>%dplyr::select(beta)), period1=c(1), period2=c(3), Index=index.names[i])
    ov3 <- data.frame(ov=overlap(temp%>%filter(period==3)%>%dplyr::select(beta),temp%>%filter(period==2)%>%dplyr::select(beta)), period1=c(3), period2=c(2), Index=index.names[i])
    #ov4 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(beta),temp%>%filter(period==3)%>%dplyr::select(beta)), period1=c(4), period2=c(3), Index=index.names[i])
   # ov5 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(beta),temp%>%filter(period==1)%>%dplyr::select(beta)), period1=c(4), period2=c(2), Index=index.names[i])
   # ov6 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(beta),temp%>%filter(period==2)%>%dplyr::select(beta)), period1=c(4), period2=c(1), Index=index.names[i])

    temp2<-rbind(ov1,ov2,ov3, ov4, ov5, ov6)
    overlap.RREAS <-rbind(temp2,overlap.RREAS)
  }
overlap.RREAS.beta<-overlap.RREAS%>%mutate(Survey="RREAS", region="CCC", season="Spring", offset=0) 


for(i in 1:4){
  temp <- RREAS%>%filter(Index==index.names[i])%>%dplyr::select(alpha, period)
  ov1 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(alpha),temp%>%filter(period==2)%>%dplyr::select(alpha)), period1=c(1), period2=c(2), Index=index.names[i])
  ov2 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(alpha),temp%>%filter(period==3)%>%dplyr::select(alpha)), period1=c(1), period2=c(3), Index=index.names[i])
  ov3 <- data.frame(ov=overlap(temp%>%filter(period==3)%>%dplyr::select(alpha),temp%>%filter(period==2)%>%dplyr::select(alpha)), period1=c(3), period2=c(2), Index=index.names[i])
 # ov4 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(alpha),temp%>%filter(period==3)%>%dplyr::select(alpha)), period1=c(4), period2=c(3), Index=index.names[i])
 # ov5 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(alpha),temp%>%filter(period==1)%>%dplyr::select(alpha)), period1=c(4), period2=c(2), Index=index.names[i])
 # ov6 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(alpha),temp%>%filter(period==2)%>%dplyr::select(alpha)), period1=c(4), period2=c(1), Index=index.names[i])
  
  temp2<-rbind(ov1,ov2,ov3, ov4, ov5, ov6)
  overlap.RREAS <-rbind(temp2,overlap.RREAS)
}
overlap.RREAS.alpha<-overlap.RREAS%>%mutate(Survey="RREAS", region="CCC", season="Spring", offset=0) 

#running models with 1-year offset

RREASoffset <-NULL
columns<-c(which(colnames(data) == "seasonal_PDO"), which(colnames(data) == "seasonal_NPGO"),
           which(colnames(data) == "seasonal_NPH"),which(colnames(data) == "seasonal_ONI"))
for(i in 1:length(columns)){
  bayeslinmod(data, columns[i],  data$estimateoffset1, data$period)
  RREASoffset <-rbind(RREASoffset,scaled.anom)
}
RREASoffset<-RREASoffset%>%mutate(survey="RREAS",
                          Index=ifelse(Index=="seasonal_PDO", "PDO", ifelse(Index=="seasonal_NPGO", "NPGO",
                                                                            ifelse(Index=="seasonal_NPH", "NPH","ONI"))))
  
postplot(RREASoffset, RREASoffset$beta)
postplot(RREASoffset, RREASoffset$alpha)

RREASoffset_full <-NULL
columns<-c(which(colnames(data) == "seasonal_PDO"), which(colnames(data) == "seasonal_NPGO"),
           which(colnames(data) == "seasonal_NPH"),which(colnames(data) == "seasonal_ONI"))
for(i in 1:length(columns)){
  bayeslinmod(data, columns[i],  data$estimate, data$full)
  RREASoffset_full <-rbind(RREASoffset_full,scaled.anom)
}

RREASoffset_full<-RREASoffset_full%>%mutate(survey="RREAS",
Index=ifelse(Index=="seasonal_PDO", "PDO", ifelse(Index=="seasonal_NPGO", "NPGO",
                                                  ifelse(Index=="seasonal_NPH", "NPH","ONI"))))%>%
  mutate(period=4)

ggplot(RREASoffset_full, aes(x = beta),fill = as.factor(period), group=as.factor(period)) +
    theme_bw() +
    facet_wrap(.~Index, ncol = 4, scales='free') +
    geom_density(alpha = 0.7) +
    scale_fill_manual(values = c(col[1])) +
    #theme(legend.title = element_blank(), legend.position = 'top', legend.key.size = unit(3, 'mm')) +
    geom_vline(xintercept = 0, lty = 2) +
    labs(x = "Slope",
         y = "Posterior density")


RREASoffset<- bind_rows(RREASoffset,RREASoffset_full)%>%
  mutate(region="CCC",Season="Spring", lag = 1)

index.names <- unique(RREASoffset$Index)
period.names <- unique(RREASoffset$period)
overlap.RREASoffset <- NA
  for(i in 1:4){
    temp <- RREASoffset%>%filter(Index==index.names[i])%>%dplyr::select(beta, period)
    ov1 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(beta),temp%>%filter(period==2)%>%dplyr::select(beta)), period1=c(1), period2=c(2), Index=index.names[i])
    ov2 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(beta),temp%>%filter(period==3)%>%dplyr::select(beta)), period1=c(1), period2=c(3), Index=index.names[i])
    ov3 <- data.frame(ov=overlap(temp%>%filter(period==3)%>%dplyr::select(beta),temp%>%filter(period==2)%>%dplyr::select(beta)), period1=c(3), period2=c(2), Index=index.names[i])
    ov4 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(beta),temp%>%filter(period==3)%>%dplyr::select(beta)), period1=c(4), period2=c(3), Index=index.names[i])
    ov5 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(beta),temp%>%filter(period==1)%>%dplyr::select(beta)), period1=c(4), period2=c(2), Index=index.names[i])
    ov6 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(beta),temp%>%filter(period==2)%>%dplyr::select(beta)), period1=c(4), period2=c(1), Index=index.names[i])

    temp2<-rbind(ov1,ov2,ov3, ov4, ov5, ov6)
    overlap.RREASoffset <-rbind(temp2,overlap.RREASoffset)
  }
overlap.RREASoffset.beta<-overlap.RREASoffset%>%mutate(Survey="RREAS", region="CCC", season="Spring", offset=1) 

overlap.RREASoffset <- NA
for(i in 1:4){
  temp <- RREASoffset%>%filter(Index==index.names[i])%>%dplyr::select(alpha, period)
  ov1 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(alpha),temp%>%filter(period==2)%>%dplyr::select(alpha)), period1=c(1), period2=c(2), Index=index.names[i])
  ov2 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(alpha),temp%>%filter(period==3)%>%dplyr::select(alpha)), period1=c(1), period2=c(3), Index=index.names[i])
  ov3 <- data.frame(ov=overlap(temp%>%filter(period==3)%>%dplyr::select(alpha),temp%>%filter(period==2)%>%dplyr::select(alpha)), period1=c(3), period2=c(2), Index=index.names[i])
  ov4 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(alpha),temp%>%filter(period==3)%>%dplyr::select(alpha)), period1=c(4), period2=c(3), Index=index.names[i])
  ov5 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(alpha),temp%>%filter(period==1)%>%dplyr::select(alpha)), period1=c(4), period2=c(2), Index=index.names[i])
  ov6 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(alpha),temp%>%filter(period==2)%>%dplyr::select(alpha)), period1=c(4), period2=c(1), Index=index.names[i])
  
  temp2<-rbind(ov1,ov2,ov3, ov4, ov5, ov6)
  overlap.RREASoffset <-rbind(temp2,overlap.RREASoffset)
}
overlap.RREASoffset.alpha<-overlap.RREASoffset%>%mutate(Survey="RREAS", region="CCC", season="Spring", offset=1) 

###### Winter #####

data <- dat%>%filter(season=="Winter", trend=="RREAS")%>%
  distinct()%>%
  filter(Year_lag<2023)%>%
  add_column(full=1)

RREAS_W <-NULL
columns<-c(which(colnames(data) == "seasonal_PDO"), which(colnames(data) == "seasonal_NPGO"),
           which(colnames(data) == "seasonal_NPH"),which(colnames(data) == "seasonal_ONI"))
for(i in 1:length(columns)){
  bayeslinmod(data, columns[i],  data$estimate, data$period)
  RREAS_W <-rbind(RREAS_W,scaled.anom)
}

RREAS_W<-RREAS_W%>%mutate(survey="RREAS",
Index=ifelse(Index=="seasonal_PDO", "PDO", ifelse(Index=="seasonal_NPGO", "NPGO",
                                                  ifelse(Index=="seasonal_NPH", "NPH","ONI"))))

postplot(RREAS_W, RREAS_W$beta)
postplot(RREAS_W, RREAS_W$alpha)

RREAS_W_full <-NULL
columns<-c(which(colnames(data) == "seasonal_PDO"), which(colnames(data) == "seasonal_NPGO"),
           which(colnames(data) == "seasonal_NPH"),which(colnames(data) == "seasonal_ONI"))
for(i in 1:length(columns)){
  bayeslinmod(data, columns[i],  data$estimate, data$full)
  RREAS_W_full <-rbind(RREAS_W_full,scaled.anom)
}

RREAS_W_full<-RREAS_W_full%>%mutate(survey="RREAS",
Index=ifelse(Index=="seasonal_PDO", "PDO", ifelse(Index=="seasonal_NPGO", "NPGO",
                                                  ifelse(Index=="seasonal_NPH", "NPH","ONI"))))%>%
  mutate(period=4)

ggplot(RREAS_W_full, aes(x = beta),fill = as.factor(period), group=as.factor(period)) +
    theme_bw() +
    facet_wrap(.~Index, ncol = 4, scales='free') +
    geom_density(alpha = 0.7) +
    scale_fill_manual(values = c(col[1])) +
    #theme(legend.title = element_blank(), legend.position = 'top', legend.key.size = unit(3, 'mm')) +
    geom_vline(xintercept = 0, lty = 2) +
    labs(x = "Slope",
         y = "Posterior density")


RREAS_W<- bind_rows(RREAS_W,RREAS_W_full)%>%
  mutate(region="CCC",Season="Winter", lag = 0)

index.names <- unique(RREAS_W$Index)
period.names <- unique(RREAS_W$period)
overlap.RREAS_W <- NA
 for(i in 1:4){
    temp <- RREAS_W%>%filter(Index==index.names[i])%>%dplyr::select(beta, period)
    ov1 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(beta),temp%>%filter(period==2)%>%dplyr::select(beta)), period1=c(1), period2=c(2), Index=index.names[i])
    ov2 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(beta),temp%>%filter(period==3)%>%dplyr::select(beta)), period1=c(1), period2=c(3), Index=index.names[i])
    ov3 <- data.frame(ov=overlap(temp%>%filter(period==3)%>%dplyr::select(beta),temp%>%filter(period==2)%>%dplyr::select(beta)), period1=c(3), period2=c(2), Index=index.names[i])
    ov4 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(beta),temp%>%filter(period==3)%>%dplyr::select(beta)), period1=c(4), period2=c(3), Index=index.names[i])
    ov5 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(beta),temp%>%filter(period==1)%>%dplyr::select(beta)), period1=c(4), period2=c(2), Index=index.names[i])
    ov6 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(beta),temp%>%filter(period==2)%>%dplyr::select(beta)), period1=c(4), period2=c(1), Index=index.names[i])

    temp2<-rbind(ov1,ov2,ov3, ov4, ov5, ov6)
    overlap.RREAS_W <-rbind(temp2,overlap.RREAS_W)
  }
overlap.RREAS_W.beta<-overlap.RREAS_W%>%mutate(Survey="RREAS", region="CCC", season="Winter", offset=0) 

overlap.RREAS_W <- NA
for(i in 1:4){
  temp <- RREAS_W%>%filter(Index==index.names[i])%>%dplyr::select(alpha, period)
  ov1 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(alpha),temp%>%filter(period==2)%>%dplyr::select(alpha)), period1=c(1), period2=c(2), Index=index.names[i])
  ov2 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(alpha),temp%>%filter(period==3)%>%dplyr::select(alpha)), period1=c(1), period2=c(3), Index=index.names[i])
  ov3 <- data.frame(ov=overlap(temp%>%filter(period==3)%>%dplyr::select(alpha),temp%>%filter(period==2)%>%dplyr::select(alpha)), period1=c(3), period2=c(2), Index=index.names[i])
  ov4 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(alpha),temp%>%filter(period==3)%>%dplyr::select(alpha)), period1=c(4), period2=c(3), Index=index.names[i])
  ov5 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(alpha),temp%>%filter(period==1)%>%dplyr::select(alpha)), period1=c(4), period2=c(2), Index=index.names[i])
  ov6 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(alpha),temp%>%filter(period==2)%>%dplyr::select(alpha)), period1=c(4), period2=c(1), Index=index.names[i])
  
  temp2<-rbind(ov1,ov2,ov3, ov4, ov5, ov6)
  overlap.RREAS_W <-rbind(temp2,overlap.RREAS_W)
}
overlap.RREAS_W.alpha<-overlap.RREAS_W%>%mutate(Survey="RREAS", region="CCC", season="Winter", offset=0) 

#running models with 1-year offset

RREAS_Woffset <-NULL
columns<-c(which(colnames(data) == "seasonal_PDO"), which(colnames(data) == "seasonal_NPGO"),
           which(colnames(data) == "seasonal_NPH"),which(colnames(data) == "seasonal_ONI"))
for(i in 1:length(columns)){
  bayeslinmod(data, columns[i],  data$estimateoffset1, data$period)
  RREAS_Woffset <-rbind(RREAS_Woffset,scaled.anom)
}
RREAS_Woffset<-RREAS_Woffset%>%mutate(survey="RREAS",
                          Index=ifelse(Index=="seasonal_PDO", "PDO", ifelse(Index=="seasonal_NPGO", "NPGO",
                                                                            ifelse(Index=="seasonal_NPH", "NPH","ONI"))))
  
postplot(RREAS_Woffset, RREAS_Woffset$beta)
postplot(RREAS_Woffset, RREAS_Woffset$alpha)

RREAS_Woffset_full <-NULL
columns<-c(which(colnames(data) == "seasonal_PDO"), which(colnames(data) == "seasonal_NPGO"),
           which(colnames(data) == "seasonal_NPH"),which(colnames(data) == "seasonal_ONI"))
for(i in 1:length(columns)){
  bayeslinmod(data, columns[i],  data$estimate, data$full)
  RREAS_Woffset_full <-rbind(RREAS_Woffset_full,scaled.anom)
}

RREAS_Woffset_full<-RREAS_Woffset_full%>%mutate(survey="RREAS",
Index=ifelse(Index=="seasonal_PDO", "PDO", ifelse(Index=="seasonal_NPGO", "NPGO",
                                                  ifelse(Index=="seasonal_NPH", "NPH","ONI"))))%>%
  mutate(period=4)

ggplot(RREAS_Woffset_full, aes(x = beta),fill = as.factor(period), group=as.factor(period)) +
    theme_bw() +
    facet_wrap(.~Index, ncol = 4, scales='free') +
    geom_density(alpha = 0.7) +
    scale_fill_manual(values = c(col[1])) +
    #theme(legend.title = element_blank(), legend.position = 'top', legend.key.size = unit(3, 'mm')) +
    geom_vline(xintercept = 0, lty = 2) +
    labs(x = "Slope",
         y = "Posterior density")


RREAS_Woffset<- bind_rows(RREAS_Woffset,RREAS_Woffset_full)%>%
  mutate(region="CCC",Season="Winter", lag = 1)

index.names <- unique(RREAS_Woffset$Index)
period.names <- unique(RREAS_Woffset$period)
overlap.RREAS_Woffset <- NA
  for(i in 1:4){
    temp <- RREAS_Woffset%>%filter(Index==index.names[i])%>%dplyr::select(beta, period)
    ov1 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(beta),temp%>%filter(period==2)%>%dplyr::select(beta)), period1=c(1), period2=c(2), Index=index.names[i])
    ov2 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(beta),temp%>%filter(period==3)%>%dplyr::select(beta)), period1=c(1), period2=c(3), Index=index.names[i])
    ov3 <- data.frame(ov=overlap(temp%>%filter(period==3)%>%dplyr::select(beta),temp%>%filter(period==2)%>%dplyr::select(beta)), period1=c(3), period2=c(2), Index=index.names[i])
    ov4 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(beta),temp%>%filter(period==3)%>%dplyr::select(beta)), period1=c(4), period2=c(3), Index=index.names[i])
    ov5 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(beta),temp%>%filter(period==1)%>%dplyr::select(beta)), period1=c(4), period2=c(2), Index=index.names[i])
    ov6 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(beta),temp%>%filter(period==2)%>%dplyr::select(beta)), period1=c(4), period2=c(1), Index=index.names[i])

    temp2<-rbind(ov1,ov2,ov3, ov4, ov5, ov6)
    overlap.RREAS_Woffset <-rbind(temp2,overlap.RREAS_Woffset)
  }
overlap.RREAS_Woffset.beta<-overlap.RREAS_Woffset%>%mutate(Survey="RREAS", region="CCC", season="Winter", offset=1) 

overlap.RREAS_Woffset <- NA
for(i in 1:4){
  temp <- RREAS_Woffset%>%filter(Index==index.names[i])%>%dplyr::select(alpha, period)
  ov1 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(alpha),temp%>%filter(period==2)%>%dplyr::select(alpha)), period1=c(1), period2=c(2), Index=index.names[i])
  ov2 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(alpha),temp%>%filter(period==3)%>%dplyr::select(alpha)), period1=c(1), period2=c(3), Index=index.names[i])
  ov3 <- data.frame(ov=overlap(temp%>%filter(period==3)%>%dplyr::select(alpha),temp%>%filter(period==2)%>%dplyr::select(alpha)), period1=c(3), period2=c(2), Index=index.names[i])
  ov4 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(alpha),temp%>%filter(period==3)%>%dplyr::select(alpha)), period1=c(4), period2=c(3), Index=index.names[i])
  ov5 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(alpha),temp%>%filter(period==1)%>%dplyr::select(alpha)), period1=c(4), period2=c(2), Index=index.names[i])
  ov6 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(alpha),temp%>%filter(period==2)%>%dplyr::select(alpha)), period1=c(4), period2=c(1), Index=index.names[i])
  
  temp2<-rbind(ov1,ov2,ov3, ov4, ov5, ov6)
  overlap.RREAS_Woffset <-rbind(temp2,overlap.RREAS_Woffset)
}
overlap.RREAS_Woffset.alpha<-overlap.RREAS_Woffset%>%mutate(Survey="RREAS", region="CCC", season="Winter", offset=1) 

#### Copepod ####
###### Spring ####
climpdo <- climate_dat_cop%>%filter(season=="Spring")%>%
  dplyr::select(Year_lag, region, seasonal_PDO,  seasonal_NPGO,  seasonal_ONI,  seasonal_NPH)%>%
  distinct()%>%
  filter(Year_lag<2023)

data <- climate_dat_cop%>%filter(season=="Summer")%>%
  dplyr::select(Year_lag, region, seasonal_copepod_northern, 
         seasonal_copepod_southern)%>%
  left_join(climpdo)%>%
  distinct()%>%
  mutate(period=ifelse(Year_lag<2013,2,3))%>%
  filter(Year_lag>1996)%>%
  mutate(full=1)

northern <-NULL
columns<-c(which(colnames(data) == "seasonal_PDO"), which(colnames(data) == "seasonal_NPGO"),
           which(colnames(data) == "seasonal_NPH"),which(colnames(data) == "seasonal_ONI"))
for(i in 1:length(columns)){
  bayeslinmod(data, columns[i],  data$seasonal_copepod_northern, data$period)
  northern <-rbind(northern,scaled.anom)
}
northern<-northern%>%mutate(survey="N. Copepod",
                            Index=ifelse(Index=="seasonal_PDO", "PDO", ifelse(Index=="seasonal_NPGO", "NPGO",
                                                                              ifelse(Index=="seasonal_NPH", "NPH","ONI"))))

northern_full <-NULL
columns<-c(which(colnames(data) == "seasonal_PDO"), which(colnames(data) == "seasonal_NPGO"),
           which(colnames(data) == "seasonal_NPH"),which(colnames(data) == "seasonal_ONI"))
for(i in 1:length(columns)){
  bayeslinmod(data, columns[i],  data$seasonal_copepod_northern, data$full)
  northern_full <-rbind(northern_full,scaled.anom)
}

northern_full<-northern_full%>%mutate(survey="N. Copepod",
Index=ifelse(Index=="seasonal_PDO", "PDO", ifelse(Index=="seasonal_NPGO", "NPGO",
                                                  ifelse(Index=="seasonal_NPH", "NPH","ONI"))))%>%
  mutate(period=4)

index.names <- unique(northern$Index)
period.names <- unique(northern$period)


northern<- bind_rows(northern,northern_full)%>%
  mutate(region="NCC",Season="Spring", lag = 0)
overlap.northern <- NA
for(i in 1:4){
  temp <- northern%>%filter(Index==index.names[i])%>%dplyr::select(beta, period)
  ov3 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(beta),temp%>%filter(period==2)%>%dplyr::select(beta)), period1=c(1), period2=c(2), Index=index.names[i])
  ov2 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(beta),temp%>%filter(period==1)%>%dplyr::select(beta)), period1=c(4), period2=c(1), Index=index.names[i])
  ov1 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(beta),temp%>%filter(period==2)%>%dplyr::select(beta)), period1=c(4), period2=c(2), Index=index.names[i])
  temp2<-rbind(ov3, ov2, ov1)
  overlap.northern <-rbind(temp2,overlap.northern)
}
overlap.northern.beta<-overlap.northern%>%mutate(Survey="N. Copepods", region="NCC", season="Spring", offset=0) 

overlap.northern <- NA
i<-1
for(i in 1:4){
  temp <- northern%>%filter(Index==index.names[i])%>%dplyr::select(alpha, period)
  ov3 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(alpha),temp%>%filter(period==2)%>%dplyr::select(alpha)), period1=c(1), period2=c(2), Index=index.names[i])
  ov2 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(alpha),temp%>%filter(period==1)%>%dplyr::select(alpha)), period1=c(4), period2=c(1), Index=index.names[i])
  ov1 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(alpha),temp%>%filter(period==2)%>%dplyr::select(alpha)), period1=c(4), period2=c(2), Index=index.names[i])
  temp2<-rbind(ov3, ov2, ov1)
  overlap.northern <-rbind(temp2,overlap.northern)
}
overlap.northern.alpha<-overlap.northern%>%mutate(Survey="N. Copepods", region="NCC", season="Spring", offset=0) 


southern <-NULL
columns<-c(which(colnames(data) == "seasonal_PDO"), which(colnames(data) == "seasonal_NPGO"),
           which(colnames(data) == "seasonal_NPH"),which(colnames(data) == "seasonal_ONI"))
for(i in 1:length(columns)){
  bayeslinmod(data, columns[i],  data$seasonal_copepod_southern, data$period)
  southern <-rbind(southern,scaled.anom)
}
southern<-southern%>%mutate(survey="Southern Copepod (NCC)",
                            Index=ifelse(Index=="seasonal_PDO", "PDO", ifelse(Index=="seasonal_NPGO", "NPGO",
                                            ifelse(Index=="seasonal_NPH", "NPH","ONI"))))

southern_full <-NULL
columns<-c(which(colnames(data) == "seasonal_PDO"), which(colnames(data) == "seasonal_NPGO"),
           which(colnames(data) == "seasonal_NPH"),which(colnames(data) == "seasonal_ONI"))
for(i in 1:length(columns)){
  bayeslinmod(data, columns[i],  data$seasonal_copepod_southern, data$full)
  southern_full <-rbind(southern_full,scaled.anom)
}

southern_full<-southern_full%>%mutate(survey="S. Copepod (NCC)",
Index=ifelse(Index=="seasonal_PDO", "PDO", ifelse(Index=="seasonal_NPGO", "NPGO",
                                                  ifelse(Index=="seasonal_NPH", "NPH","ONI"))))%>%
  mutate(period=4)

index.names <- unique(southern$Index)
period.names <- unique(southern$period)
southern<- bind_rows(southern,southern_full)%>%
  mutate(region="NCC",Season="Spring", lag = 0)
overlap.southern <- NA

for(i in 1:4){
  temp <- southern%>%filter(Index==index.names[i])%>%dplyr::select(beta, period)
  ov3 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(beta),temp%>%filter(period==2)%>%dplyr::select(beta)), period1=c(1), period2=c(2), Index=index.names[i])
  ov2 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(beta),temp%>%filter(period==1)%>%dplyr::select(beta)), period1=c(4), period2=c(1), Index=index.names[i])
  ov1 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(beta),temp%>%filter(period==2)%>%dplyr::select(beta)), period1=c(4), period2=c(2), Index=index.names[i])
  temp2<-rbind(ov3, ov2, ov1)
  overlap.southern <-rbind(temp2,overlap.southern)
}
overlap.southern.beta<-overlap.southern%>%mutate(Survey="S. Copepods", region="NCC", season="Spring", offset=0) 

overlap.southern <- NA
for(i in 1:4){
  temp <- southern%>%filter(Index==index.names[i])%>%dplyr::select(alpha, period)
  ov3 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(alpha),temp%>%filter(period==2)%>%dplyr::select(alpha)), period1=c(1), period2=c(2), Index=index.names[i])
  ov2 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(alpha),temp%>%filter(period==1)%>%dplyr::select(alpha)), period1=c(4), period2=c(1), Index=index.names[i])
  ov1 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(alpha),temp%>%filter(period==2)%>%dplyr::select(alpha)), period1=c(4), period2=c(2), Index=index.names[i])
  temp2<-rbind(ov3, ov2, ov1)
  overlap.southern <-rbind(temp2,overlap.southern)
}
overlap.southern.alpha<-overlap.southern%>%mutate(Survey="S. Copepods", region="NCC", season="Spring", offset=0) 

###### Winter ####
climpdo <- climate_dat_cop%>%filter(season=="Winter")%>%
  dplyr::select(Year_lag, region, seasonal_PDO,  seasonal_NPGO,  seasonal_ONI,  seasonal_NPH)%>%
  distinct()%>%
  filter(Year_lag<2023)

data <- climate_dat_cop%>%filter(season=="Summer")%>%
  dplyr::select(Year_lag, region, seasonal_copepod_northern, 
         seasonal_copepod_southern)%>%
  left_join(climpdo)%>%
  distinct()%>%
  mutate(period=ifelse(Year_lag<2013,2,3))%>%
  filter(Year_lag>1996)%>%
  mutate(full=1)

northernW <-NULL
columns<-c(which(colnames(data) == "seasonal_PDO"), which(colnames(data) == "seasonal_NPGO"),
           which(colnames(data) == "seasonal_NPH"),which(colnames(data) == "seasonal_ONI"))
for(i in 1:length(columns)){
  bayeslinmod(data, columns[i],  data$seasonal_copepod_northern, data$period)
  northernW <-rbind(northernW,scaled.anom)
}
northernW<-northernW%>%mutate(survey="N. Copepod (NCC)",
                            Index=ifelse(Index=="seasonal_PDO", "PDO", ifelse(Index=="seasonal_NPGO", "NPGO",
                                                                              ifelse(Index=="seasonal_NPH", "NPH","ONI"))))

northernW_full <-NULL
columns<-c(which(colnames(data) == "seasonal_PDO"), which(colnames(data) == "seasonal_NPGO"),
           which(colnames(data) == "seasonal_NPH"),which(colnames(data) == "seasonal_ONI"))
for(i in 1:length(columns)){
  bayeslinmod(data, columns[i],  data$seasonal_copepod_northern, data$full)
  northernW_full <-rbind(northernW_full,scaled.anom)
}

northernW_full<-northernW_full%>%mutate(survey="N. Copepod (NCC)",
Index=ifelse(Index=="seasonal_PDO", "PDO", ifelse(Index=="seasonal_NPGO", "NPGO",
                                                  ifelse(Index=="seasonal_NPH", "NPH","ONI"))))%>%
  mutate(period=4)

index.names <- unique(northernW$Index)
period.names <- unique(northernW$period)


northernW<- bind_rows(northernW,northernW_full)%>%
  mutate(region="NCC",Season="Winter", lag = 0)
overlap.northernW <- NA
for(i in 1:4){
  temp <- northernW%>%filter(Index==index.names[i])%>%dplyr::select(beta, period)
  ov3 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(beta),temp%>%filter(period==2)%>%dplyr::select(beta)), period1=c(1), period2=c(2), Index=index.names[i])
  ov2 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(beta),temp%>%filter(period==1)%>%dplyr::select(beta)), period1=c(4), period2=c(1), Index=index.names[i])
  ov1 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(beta),temp%>%filter(period==2)%>%dplyr::select(beta)), period1=c(4), period2=c(2), Index=index.names[i])
  temp2<-rbind(ov3, ov2, ov1)
  overlap.northernW <-rbind(temp2,overlap.northernW)
}
overlap.northernW.beta<-overlap.northernW%>%mutate(Survey="N. Copepods", region="NCC", season="Winter", offset=0) 

overlap.northernW <- NA
for(i in 1:4){
  temp <- northernW%>%filter(Index==index.names[i])%>%dplyr::select(alpha, period)
  ov3 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(alpha),temp%>%filter(period==2)%>%dplyr::select(alpha)), period1=c(1), period2=c(2), Index=index.names[i])
  ov2 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(alpha),temp%>%filter(period==1)%>%dplyr::select(alpha)), period1=c(4), period2=c(1), Index=index.names[i])
  ov1 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(alpha),temp%>%filter(period==2)%>%dplyr::select(alpha)), period1=c(4), period2=c(2), Index=index.names[i])
  temp2<-rbind(ov3, ov2, ov1)
  overlap.northernW <-rbind(temp2,overlap.northernW)
}
overlap.northernW.alpha<-overlap.northernW%>%mutate(Survey="N. Copepods", region="NCC", season="Winter", offset=0) 


southernW <-NULL
columns<-c(which(colnames(data) == "seasonal_PDO"), which(colnames(data) == "seasonal_NPGO"),
           which(colnames(data) == "seasonal_NPH"),which(colnames(data) == "seasonal_ONI"))
for(i in 1:length(columns)){
  bayeslinmod(data, columns[i],  data$seasonal_copepod_southern, data$period)
  southernW <-rbind(southernW,scaled.anom)
}
southernW<-southernW%>%mutate(survey="S. Copepod (NCC)",
                            Index=ifelse(Index=="seasonal_PDO", "PDO", ifelse(Index=="seasonal_NPGO", "NPGO",
                                                                              ifelse(Index=="seasonal_NPH", "NPH","ONI"))))

southernW_full <-NULL
columns<-c(which(colnames(data) == "seasonal_PDO"), which(colnames(data) == "seasonal_NPGO"),
           which(colnames(data) == "seasonal_NPH"),which(colnames(data) == "seasonal_ONI"))
for(i in 1:length(columns)){
  bayeslinmod(data, columns[i],  data$seasonal_copepod_southern, data$full)
  southernW_full <-rbind(southernW_full,scaled.anom)
}

southernW_full<-southernW_full%>%mutate(survey="S. Copepod (NCC)",
Index=ifelse(Index=="seasonal_PDO", "PDO", ifelse(Index=="seasonal_NPGO", "NPGO",
                                                  ifelse(Index=="seasonal_NPH", "NPH","ONI"))))%>%
  mutate(period=4)

index.names <- unique(southernW$Index)
period.names <- unique(southernW$period)
southernW<- bind_rows(southernW,southernW_full)%>%
  mutate(region="NCC",Season="Winter", lag = 0)
overlap.southernW <- NA
for(i in 1:4){
  temp <- southernW%>%filter(Index==index.names[i])%>%dplyr::select(beta, period)
  ov3 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(beta),temp%>%filter(period==2)%>%dplyr::select(beta)), period1=c(1), period2=c(2), Index=index.names[i])
  ov2 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(beta),temp%>%filter(period==1)%>%dplyr::select(beta)), period1=c(4), period2=c(1), Index=index.names[i])
  ov1 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(beta),temp%>%filter(period==2)%>%dplyr::select(beta)), period1=c(4), period2=c(2), Index=index.names[i])
  temp2<-rbind(ov3, ov2, ov1)
  overlap.southernW <-rbind(temp2,overlap.southernW)
}
overlap.southernW.beta<-overlap.southernW%>%mutate(Survey="S. Copepods", region="NCC", season="Winter", offset=0) 


overlap.southernW <- NA
for(i in 1:4){
  temp <- southernW%>%filter(Index==index.names[i])%>%dplyr::select(alpha, period)
  ov3 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(alpha),temp%>%filter(period==2)%>%dplyr::select(alpha)), period1=c(1), period2=c(2), Index=index.names[i])
  ov2 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(alpha),temp%>%filter(period==1)%>%dplyr::select(alpha)), period1=c(4), period2=c(1), Index=index.names[i])
  ov1 <- data.frame(ov=overlap(temp%>%filter(period==4)%>%dplyr::select(alpha),temp%>%filter(period==2)%>%dplyr::select(alpha)), period1=c(4), period2=c(2), Index=index.names[i])
  temp2<-rbind(ov3, ov2, ov1)
  overlap.southernW <-rbind(temp2,overlap.southernW)
}
overlap.southernW.alpha<-overlap.southernW%>%mutate(Survey="S. Copepods", region="NCC", season="Winter", offset=0) 


#### Upwelling Model Runs ####

###### spring #####
climate_dat <-readRDS(here('data/physical/climate_dat_upwelling.rds'))
season <- "Spring"
data<- climate_dat%>%filter(season=="Spring")%>%
  #  filter(Year_lag!=2022&Year_lag!=2015)%>%
  distinct()

upwelling <-NULL
columns<-c(which(colnames(data) == "seasonal_PDO"), which(colnames(data) == "seasonal_NPGO"),
           which(colnames(data) == "seasonal_NPH"),which(colnames(data) == "seasonal_ONI"))
for(i in 1:length(columns)){
  bayeslinmod(data, columns[i],  data$stand_bakun_seasonally, data$era.region)
  upwelling <-rbind(upwelling,scaled.anom)
}

upwelling<-upwelling%>%mutate(survey="Upwelling",era.region=period,
                            Index=ifelse(Index=="seasonal_PDO", "PDO", ifelse(Index=="seasonal_NPGO", "NPGO",
                                ifelse(Index=="seasonal_NPH", "NPH","ONI"))))%>%
  dplyr::select(!period)%>%
  left_join(data%>%dplyr::select(region,era.region,period)%>%distinct())%>%
  mutate(Season='Spring', lag=0)


unique(upwelling$period)



upwelling_full <-NULL
columns<-c(which(colnames(data) == "seasonal_PDO"), which(colnames(data) == "seasonal_NPGO"),
           which(colnames(data) == "seasonal_NPH"),which(colnames(data) == "seasonal_ONI"))
for(i in 1:length(columns)){
  bayeslinmod(data, columns[i],  data$stand_bakun_seasonally, data$region)
  upwelling_full <-rbind(upwelling_full,scaled.anom)
}

reg<-unique(data%>%dplyr::select(region))%>%
  rename(region2=region)%>%
  mutate(region=as.numeric(as.factor(region2)))

upwelling_full<-upwelling_full%>%mutate(survey="Upwelling",region=period,
                            Index=ifelse(Index=="seasonal_PDO", "PDO", ifelse(Index=="seasonal_NPGO", "NPGO",
                                ifelse(Index=="seasonal_NPH", "NPH","ONI"))))%>%
  left_join(reg)%>%mutate(period=4)%>%dplyr::select(!region)%>%rename(region=region2)%>%
  mutate(Season='Spring', lag=0)


upwelling <- upwelling%>%dplyr::select(!era.region)%>%mutate(period=as.numeric(as.factor(period)))%>%
  bind_rows(upwelling_full%>%dplyr::select(alpha,beta,Index,yfirst,ylast,survey,region,period,Season,lag))

index.names <- unique(upwelling$Index)
region.names<-unique(upwelling$region)
period.names <- unique(upwelling$period)
overlap.up <- NA
for(j in 1:4){
  temp1 <- upwelling%>%filter(region==region.names[j])
  for(i in 1:4){
    temp <- temp1%>%filter(Index==index.names[i])%>%dplyr::select(beta, period,region)
    ov1 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(beta),temp%>%filter(period==2)%>%dplyr::select(beta)), period1=c(1), period2=c(2), Index=index.names[i], region=region.names[j])
    ov2 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(beta),temp%>%filter(period==3)%>%dplyr::select(beta)), period1=c(1), period2=c(3), Index=index.names[i], region=region.names[j])
    ov3 <- data.frame(ov=overlap(temp%>%filter(period==3)%>%dplyr::select(beta),temp%>%filter(period==2)%>%dplyr::select(beta)), period1=c(3), period2=c(2), Index=index.names[i], region=region.names[j])
    ov4 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(beta),temp%>%filter(period==4)%>%dplyr::select(beta)), period1=c(1), period2=c(4), Index=index.names[i], region=region.names[j])
    ov5 <- data.frame(ov=overlap(temp%>%filter(period==2)%>%dplyr::select(beta),temp%>%filter(period==4)%>%dplyr::select(beta)), period1=c(2), period2=c(4), Index=index.names[i], region=region.names[j])
    ov6 <- data.frame(ov=overlap(temp%>%filter(period==3)%>%dplyr::select(beta),temp%>%filter(period==4)%>%dplyr::select(beta)), period1=c(3), period2=c(4), Index=index.names[i], region=region.names[j])

    temp2<-rbind(ov1,ov2,ov3,ov4,ov5,ov6)
    overlap.up <-rbind(temp2,overlap.up)
  }
  
}

overlap.up


#### Upwelling Model Runs ####

###### winter #####
climate_dat <-readRDS(here('data/physical/climate_dat_upwelling.rds'))
season <- "Winter"
data<- climate_dat%>%filter(season=="Winter")%>%
  #  filter(Year_lag!=2022&Year_lag!=2015)%>%
  distinct()

upwellingW <-NULL
columns<-c(which(colnames(data) == "seasonal_PDO"), which(colnames(data) == "seasonal_NPGO"),
           which(colnames(data) == "seasonal_NPH"),which(colnames(data) == "seasonal_ONI"))
for(i in 1:length(columns)){
  bayeslinmod(data, columns[i],  data$stand_bakun_seasonally, data$era.region)
  upwellingW <-rbind(upwellingW,scaled.anom)
}

upwellingW<-upwellingW%>%mutate(survey="Upwelling",era.region=period,
                            Index=ifelse(Index=="seasonal_PDO", "PDO", ifelse(Index=="seasonal_NPGO", "NPGO",
                                ifelse(Index=="seasonal_NPH", "NPH","ONI"))))%>%
  dplyr::select(!period)%>%
  left_join(data%>%dplyr::select(region,era.region,period)%>%distinct())%>%
  mutate(Season='Winter', lag=0)


unique(upwellingW$period)



upwellingW_full <-NULL
columns<-c(which(colnames(data) == "seasonal_PDO"), which(colnames(data) == "seasonal_NPGO"),
           which(colnames(data) == "seasonal_NPH"),which(colnames(data) == "seasonal_ONI"))
for(i in 1:length(columns)){
  bayeslinmod(data, columns[i],  data$stand_bakun_seasonally, data$region)
  upwellingW_full <-rbind(upwellingW_full,scaled.anom)
}

reg<-unique(data%>%dplyr::select(region))%>%
  rename(region2=region)%>%
  mutate(region=as.numeric(as.factor(region2)))

upwellingW_full<-upwellingW_full%>%mutate(survey="Upwelling",region=period,
                            Index=ifelse(Index=="seasonal_PDO", "PDO", ifelse(Index=="seasonal_NPGO", "NPGO",
                                ifelse(Index=="seasonal_NPH", "NPH","ONI"))))%>%
  left_join(reg)%>%mutate(period=4)%>%dplyr::select(!region)%>%rename(region=region2)%>%
  mutate(Season='Winter', lag=0)


upwellingW <- upwellingW%>%dplyr::select(!era.region)%>%mutate(period=as.numeric(as.factor(period)))%>%
  bind_rows(upwellingW_full%>%dplyr::select(alpha,beta,Index,yfirst,ylast,survey,region,period,Season,lag))

index.names <- unique(upwellingW$Index)
region.names<-unique(upwellingW$region)
period.names <- unique(upwellingW$period)
overlap.upW <- NA
for(j in 1:4){
  temp1 <- upwellingW%>%filter(region==region.names[j])
  for(i in 1:4){
    temp <- temp1%>%filter(Index==index.names[i])%>%dplyr::select(beta, period,region)
    ov1 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(beta),temp%>%filter(period==2)%>%dplyr::select(beta)), period1=c(1), period2=c(2), Index=index.names[i], region=region.names[j])
    ov2 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(beta),temp%>%filter(period==3)%>%dplyr::select(beta)), period1=c(1), period2=c(3), Index=index.names[i], region=region.names[j])
    ov3 <- data.frame(ov=overlap(temp%>%filter(period==3)%>%dplyr::select(beta),temp%>%filter(period==2)%>%dplyr::select(beta)), period1=c(3), period2=c(2), Index=index.names[i], region=region.names[j])
    ov4 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(beta),temp%>%filter(period==4)%>%dplyr::select(beta)), period1=c(1), period2=c(4), Index=index.names[i], region=region.names[j])
    ov5 <- data.frame(ov=overlap(temp%>%filter(period==2)%>%dplyr::select(beta),temp%>%filter(period==4)%>%dplyr::select(beta)), period1=c(2), period2=c(4), Index=index.names[i], region=region.names[j])
    ov6 <- data.frame(ov=overlap(temp%>%filter(period==3)%>%dplyr::select(beta),temp%>%filter(period==4)%>%dplyr::select(beta)), period1=c(3), period2=c(4), Index=index.names[i], region=region.names[j])

    temp2<-rbind(ov1,ov2,ov3,ov4,ov5,ov6)
    overlap.upW <-rbind(temp2,overlap.upW)
  }
  
}

overlap.upW


##### Phenology Model Runs #####
climate_dat <-readRDS(here('data/physical/climate_dat_upwelling.rds'))%>%
  mutate(region=fct_relevel(region,c("Northern CC","Central CC","Southern CC")))
tum<- climate_dat%>%filter(season=="Spring")%>%
  dplyr::select(Year_lag, region,stand_tumi,seasonal_NPGO)%>%
  distinct()%>%
  pivot_longer(-c(Year_lag, region),names_to = 'indices', values_to = 'values')
ggplot(data=tum, aes(x=Year_lag, y=values,color=indices))+
  facet_wrap(~region)+
  geom_point()+
  geom_line()
 
season <- "Spring"
eras <- data.frame(era.region2=seq(1,9), era.region=seq(4,12))
data<- climate_dat%>%filter(season=="Spring"&region!='GoA'&Year_lag<2023&period!=4)%>%
  #  filter(Year_lag!=2022&Year_lag!=2015)%>%
  distinct()%>%
  left_join(eras)
index.names <-c("PDO", "NPGO", "NPH", "ONI")
region.names<-unique(climate_dat$region)
period.names <- unique(climate_dat$period)

data.phe.lm<-data%>%filter(season=="Spring")%>%
  dplyr::select(Year_lag, season, period,region,stand_tumi,stand_lusi,stand_sti,#stand_bakun_seasonally,
                seasonal_NPH,seasonal_NPGO,seasonal_PDO,seasonal_ONI)%>%
  rename(NPH=seasonal_NPH,NPGO=seasonal_NPGO,PDO=seasonal_PDO,ONI=seasonal_ONI)%>%
  distinct()%>%
  pivot_longer(!c(Year_lag, season, period,region,stand_tumi,stand_lusi,stand_sti), 
               names_to = "Index_Name", values_to = "Index_Value")%>%
  pivot_longer(!c(Year_lag, season,period,region, Index_Value, Index_Name), 
               names_to = "Up_Name", values_to = "Up_Value")%>%
  distinct()

###### STI #######

sti_dat <-ggplot(data = data.phe.lm%>%filter(Up_Name=="stand_sti"&period!=4), aes(y = Up_Value, x =Index_Value,col=as.factor(period))) +
  facet_grid(Index_Name~region, scales='free') +
  geom_point(aes(col=as.factor(period)), alpha = 0.5) +
  # geom_text(aes(label=Year_lag,col=as.factor(period))) +
  geom_smooth(method = "lm", se = FALSE, aes(col=as.factor(period))) +
  #geom_smooth(method = "lm", se = FALSE, col='grey') +
  scale_y_continuous(name = "STI") +
  scale_color_manual(values =  col[1:3], name="Period",labels=c('1967 - 1988', '1989 - 2012','2013 - 2022'))+
  theme_bw()+
  xlab("Climate Index Value")+
  theme(plot.title = element_text(hjust = 0.5))+
  ggtitle("Spring")
sti_dat
pdf(file = "Output/FigureS10_STIlinearregression.pdf",   # The directory you want to save the file in
    width = 7, # The width of the plot in inches
    height = 5)
sti_dat
dev.off()



STI<-NULL
columns<-c(which(colnames(data) == "seasonal_PDO"), which(colnames(data) == "seasonal_NPGO"),
           which(colnames(data) == "seasonal_NPH"),which(colnames(data) == "seasonal_ONI"))
for(i in 1:length(columns)){
  bayeslinmod(data, columns[i],  data$stand_sti, data$era.region2)
  STI <-rbind(STI,scaled.anom)
}
STI<-STI%>%mutate(survey="STI",era.region2=period,
                Index=ifelse(Index=="seasonal_PDO", "PDO", ifelse(Index=="seasonal_NPGO", "NPGO",
                                                                  ifelse(Index=="seasonal_NPH", "NPH","ONI"))))%>%
  dplyr::select(!period)%>%
  left_join(data%>%dplyr::select(region,era.region2,period)%>%distinct())%>%
  mutate(Season='Spring', lag=0, era.region=era.region2,period=as.numeric(as.factor(period)))


STI_beta <-ggplot(STI, aes(x = beta, fill = as.factor(period))) +
  theme_bw() +
  facet_grid(Index~region) +
  geom_density(alpha = 0.7) +
  scale_fill_manual(values = c(col[1],col[2], col[3]), name="Period",labels=c('1967 - 1988', '1989 - 2012','2013 - 2022')) +
  #theme(legend.title = element_blank(), legend.position = 'top', legend.key.size = unit(3, 'mm')) +
  geom_vline(xintercept = 0, lty = 2) +
  labs(x = "Slope",
       y = "Posterior density")
pdf(file = "Output/FigureS11_STIbeta.pdf",   # The directory you want to save the file in
    width = 7, # The width of the plot in inches
    height = 5)
STI_beta
dev.off()

STI_alpha <-ggplot(STI, aes(x = alpha, fill = as.factor(period))) +
  theme_bw() +
  facet_grid(Index~region) +
  geom_density(alpha = 0.7) +
  scale_fill_manual(values = c(col[1],col[2], col[3]), name="Period",labels=c('1967 - 1988', '1989 - 2012','2013 - 2022')) +
  #theme(legend.title = element_blank(), legend.position = 'top', legend.key.size = unit(3, 'mm')) +
  geom_vline(xintercept = 0, lty = 2) +
  labs(x = "Intercept",
       y = "Posterior density")
pdf(file = "Output/FigureS12_STIalpha.pdf",   # The directory you want to save the file in
    width = 7, # The width of the plot in inches
    height = 5)
STI_alpha
dev.off()

index.names <-unique(STI$Index)
region.names<-unique(STI$region)
period.names <- unique(STI$period)

postplot(STI, STI$beta)
postplot(STI,STI$alpha)
as.numeric(STI$period)
overlap.sti <- NA
for(j in 1:3){
  temp1 <- STI%>%filter(region==region.names[j])
  for(i in 1:4){
    temp <- temp1%>%filter(Index==index.names[i])%>%dplyr::select(beta, period,region)
    ov1 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(beta),temp%>%filter(period==2)%>%dplyr::select(beta)), period1=c(1), period2=c(2), Index=index.names[i], region=region.names[j])
    ov2 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(beta),temp%>%filter(period==3)%>%dplyr::select(beta)), period1=c(1), period2=c(3), Index=index.names[i], region=region.names[j])
    ov3 <- data.frame(ov=overlap(temp%>%filter(period==3)%>%dplyr::select(beta),temp%>%filter(period==2)%>%dplyr::select(beta)), period1=c(3), period2=c(2), Index=index.names[i], region=region.names[j])
    temp2<-rbind(ov1,ov2,ov3)
    overlap.sti <-rbind(temp2,overlap.sti)
  }
  
}

overlap.sti.beta<-overlap.sti%>%mutate(survey="STI")

overlap.sti <- NA
for(j in 1:3){
  temp1 <- STI%>%filter(region==region.names[j])
  for(i in 1:4){
    temp <- temp1%>%filter(Index==index.names[i])%>%dplyr::select(alpha, period,region)
    ov1 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(alpha),temp%>%filter(period==2)%>%dplyr::select(alpha)), period1=c(1), period2=c(2), Index=index.names[i], region=region.names[j])
    ov2 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(alpha),temp%>%filter(period==3)%>%dplyr::select(alpha)), period1=c(1), period2=c(3), Index=index.names[i], region=region.names[j])
    ov3 <- data.frame(ov=overlap(temp%>%filter(period==3)%>%dplyr::select(alpha),temp%>%filter(period==2)%>%dplyr::select(alpha)), period1=c(3), period2=c(2), Index=index.names[i], region=region.names[j])
    temp2<-rbind(ov1,ov2,ov3)
    overlap.sti <-rbind(temp2,overlap.sti)
  }
  
}

overlap.sti.alpha<-overlap.sti%>%mutate(survey="STI")

###### LUSI #######
lusi_dat <-ggplot(data = data.phe.lm%>%filter(Up_Name=="stand_lusi"&period!=4), aes(y = Up_Value, x =Index_Value,col=as.factor(period))) +
  facet_grid(Index_Name~region, scales='free') +
  geom_point(aes(col=as.factor(period)), alpha = 0.5) +
  # geom_text(aes(label=Year_lag,col=as.factor(period))) +
  geom_smooth(method = "lm", se = FALSE, aes(col=as.factor(period))) +
  #geom_smooth(method = "lm", se = FALSE, col='grey') +
  scale_y_continuous(name = "LUSI") +
  scale_color_manual(values =  col[1:3], name="Period",labels=c('1967 - 1988', '1989 - 2012','2013 - 2022'))+
  theme_bw()+
  xlab("Climate Index Value")+
  theme(plot.title = element_text(hjust = 0.5))+
  ggtitle("Spring")
lusi_dat
pdf(file = "Output/FigureS13_LUSIlinearregression.pdf",   # The directory you want to save the file in
    width = 7, # The width of the plot in inches
    height = 5)
lusi_dat
dev.off()

LUSI<-NULL
for(i in 1:length(columns)){
  bayeslinmod(data, columns[i],  data$stand_lusi, data$era.region2)
  LUSI <-rbind(LUSI,scaled.anom)
}
LUSI<-LUSI%>%mutate(survey="LUSI",era.region2=period,
                  Index=ifelse(Index=="seasonal_PDO", "PDO", ifelse(Index=="seasonal_NPGO", "NPGO",
                                                                    ifelse(Index=="seasonal_NPH", "NPH","ONI"))))%>%
  dplyr::select(!period)%>%
  left_join(data%>%dplyr::select(region,era.region2,period)%>%distinct())%>%
  mutate(Season='Spring', lag=0, era.region=era.region2,period=as.numeric(as.factor(period)))



LUSI_beta <-ggplot(LUSI, aes(x = beta, fill = as.factor(period))) +
  theme_bw() +
  facet_grid(Index~region) +
  geom_density(alpha = 0.7) +
  scale_fill_manual(values = c(col[1],col[2], col[3]), name="Period",labels=c('1967 - 1988', '1989 - 2012','2013 - 2022')) +
  #theme(legend.title = element_blank(), legend.position = 'top', legend.key.size = unit(3, 'mm')) +
  geom_vline(xintercept = 0, lty = 2) +
  labs(x = "Slope",
       y = "Posterior density")
pdf(file = "Output/FigureS14_LUSIbeta.pdf",   # The directory you want to save the file in
    width = 7, # The width of the plot in inches
    height = 5)
LUSI_beta
dev.off()

LUSI_alpha <-ggplot(LUSI, aes(x = alpha, fill = as.factor(period))) +
  theme_bw() +
  facet_grid(Index~region) +
  geom_density(alpha = 0.7) +
  scale_fill_manual(values = c(col[1],col[2], col[3]), name="Period",labels=c('1967 - 1988', '1989 - 2012','2013 - 2022')) +
  #theme(legend.title = element_blank(), legend.position = 'top', legend.key.size = unit(3, 'mm')) +
  geom_vline(xintercept = 0, lty = 2) +
  labs(x = "Intercept",
       y = "Posterior density")
pdf(file = "Output/FigureS15_LUSIalpha.pdf",   # The directory you want to save the file in
    width = 7, # The width of the plot in inches
    height = 5)
LUSI_alpha
dev.off()

overlap.lusi <- NA
for(j in 1:3){
  temp1 <- LUSI%>%filter(region==region.names[j])
  for(i in 1:4){
    temp <- temp1%>%filter(Index==index.names[i])%>%dplyr::select(beta, period,region)
    ov1 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(beta),temp%>%filter(period==2)%>%dplyr::select(beta)), period1=c(1), period2=c(2), Index=index.names[i], region=region.names[j])
    ov2 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(beta),temp%>%filter(period==3)%>%dplyr::select(beta)), period1=c(1), period2=c(3), Index=index.names[i], region=region.names[j])
    ov3 <- data.frame(ov=overlap(temp%>%filter(period==3)%>%dplyr::select(beta),temp%>%filter(period==2)%>%dplyr::select(beta)), period1=c(3), period2=c(2), Index=index.names[i], region=region.names[j])
    temp2<-rbind(ov1,ov2,ov3)
    overlap.lusi <-rbind(temp2,overlap.lusi)
  }
  
}

overlap.lusi
overlap.lusi.beta<-overlap.lusi%>%mutate(survey="LUSI")

overlap.lusi <- NA
for(j in 1:3){
  temp1 <- LUSI%>%filter(region==region.names[j])
  for(i in 1:4){
    temp <- temp1%>%filter(Index==index.names[i])%>%dplyr::select(alpha, period,region)
    ov1 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(alpha),temp%>%filter(period==2)%>%dplyr::select(alpha)), period1=c(1), period2=c(2), Index=index.names[i], region=region.names[j])
    ov2 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(alpha),temp%>%filter(period==3)%>%dplyr::select(alpha)), period1=c(1), period2=c(3), Index=index.names[i], region=region.names[j])
    ov3 <- data.frame(ov=overlap(temp%>%filter(period==3)%>%dplyr::select(alpha),temp%>%filter(period==2)%>%dplyr::select(alpha)), period1=c(3), period2=c(2), Index=index.names[i], region=region.names[j])
    temp2<-rbind(ov1,ov2,ov3)
    overlap.lusi <-rbind(temp2,overlap.lusi)
  }
  
}

overlap.lusi
overlap.lusi.alpha<-overlap.lusi%>%mutate(survey="LUSI")

###### TUMI #######
tumi_dat <-ggplot(data = data.phe.lm%>%filter(Up_Name=="stand_tumi"&period!=4), aes(y = Up_Value, x =Index_Value,col=as.factor(period))) +
  facet_grid(Index_Name~region, scales='free') +
  geom_point(aes(col=as.factor(period)), alpha = 0.5) +
  # geom_text(aes(label=Year_lag,col=as.factor(period))) +
  geom_smooth(method = "lm", se = FALSE, aes(col=as.factor(period))) +
  #geom_smooth(method = "lm", se = FALSE, col='grey') +
  scale_y_continuous(name = "TUMI") +
  scale_color_manual(values =  col[1:3], name="Period",labels=c('1967 - 1988', '1989 - 2012','2013 - 2022'))+
  theme_bw()+
  xlab("Climate Index Value")+
  theme(plot.title = element_text(hjust = 0.5))+
  ggtitle("Spring")
tumi_dat
pdf(file = "Output/FigureS7_TUMIlinearregression.pdf",   # The directory you want to save the file in
    width = 7, # The width of the plot in inches
    height = 5)
tumi_dat
dev.off()

TUMI<-NULL
for(i in 1:length(columns)){
  bayeslinmod(data, columns[i],  data$stand_tumi, data$era.region2)
  TUMI <-rbind(TUMI,scaled.anom)
}
TUMI<-TUMI%>%mutate(survey="TUMI",era.region2=period,
                    Index=ifelse(Index=="seasonal_PDO", "PDO", ifelse(Index=="seasonal_NPGO", "NPGO",
                                                                      ifelse(Index=="seasonal_NPH", "NPH","ONI"))))%>%
  dplyr::select(!period)%>%
  left_join(data%>%dplyr::select(region,era.region2,period)%>%distinct())%>%
  mutate(Season='Spring', lag=0, era.region=era.region2,period=as.numeric(as.factor(period)))


TUMI_beta <-ggplot(TUMI, aes(x = beta, fill = as.factor(period))) +
  theme_bw() +
  facet_grid(Index~region) +
  geom_density(alpha = 0.7) +
  scale_fill_manual(values = c(col[1],col[2], col[3]), name="Period",labels=c('1967 - 1988', '1989 - 2012','2013 - 2022')) +
  #theme(legend.title = element_blank(), legend.position = 'top', legend.key.size = unit(3, 'mm')) +
  geom_vline(xintercept = 0, lty = 2) +
  labs(x = "Slope",
       y = "Posterior density")
pdf(file = "Output/FigureS8_TUMIbeta.pdf",   # The directory you want to save the file in
    width = 7, # The width of the plot in inches
    height = 5)
TUMI_beta
dev.off()

TUMI_alpha <-ggplot(TUMI, aes(x = alpha, fill = as.factor(period))) +
  theme_bw() +
  facet_grid(Index~region) +
  geom_density(alpha = 0.7) +
  scale_fill_manual(values = c(col[1],col[2], col[3]), name="Period",labels=c('1967 - 1988', '1989 - 2012','2013 - 2022')) +
  #theme(legend.title = element_blank(), legend.position = 'top', legend.key.size = unit(3, 'mm')) +
  geom_vline(xintercept = 0, lty = 2) +
  labs(x = "Intercept",
       y = "Posterior density")
pdf(file = "Output/FigureS9_TUMIalpha.pdf",   # The directory you want to save the file in
    width = 7, # The width of the plot in inches
    height = 5)
TUMI_alpha
dev.off()


overlap.tumi <- NA
for(j in 1:3){
  temp1 <- TUMI%>%filter(region==region.names[j])
  for(i in 1:4){
    temp <- temp1%>%filter(Index==index.names[i])%>%dplyr::select(beta, period,region)
    ov1 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(beta),temp%>%filter(period==2)%>%dplyr::select(beta)), period1=c(1), period2=c(2), Index=index.names[i], region=region.names[j])
    ov2 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(beta),temp%>%filter(period==3)%>%dplyr::select(beta)), period1=c(1), period2=c(3), Index=index.names[i], region=region.names[j])
    ov3 <- data.frame(ov=overlap(temp%>%filter(period==3)%>%dplyr::select(beta),temp%>%filter(period==2)%>%dplyr::select(beta)), period1=c(3), period2=c(2), Index=index.names[i], region=region.names[j])
    temp2<-rbind(ov1,ov2,ov3)
    overlap.tumi <-rbind(temp2,overlap.tumi)
  }
  
}
overlap.tumi.beta<- overlap.tumi%>%mutate(survey="TUMI")


overlap.tumi <- NA
for(j in 1:3){
  temp1 <- TUMI%>%filter(region==region.names[j])
  for(i in 1:4){
    temp <- temp1%>%filter(Index==index.names[i])%>%dplyr::select(alpha, period,region)
    ov1 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(alpha),temp%>%filter(period==2)%>%dplyr::select(alpha)), period1=c(1), period2=c(2), Index=index.names[i], region=region.names[j])
    ov2 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(alpha),temp%>%filter(period==3)%>%dplyr::select(alpha)), period1=c(1), period2=c(3), Index=index.names[i], region=region.names[j])
    ov3 <- data.frame(ov=overlap(temp%>%filter(period==3)%>%dplyr::select(alpha),temp%>%filter(period==2)%>%dplyr::select(alpha)), period1=c(3), period2=c(2), Index=index.names[i], region=region.names[j])
    temp2<-rbind(ov1,ov2,ov3)
    overlap.tumi <-rbind(temp2,overlap.tumi)
  }
  
}
overlap.tumi.alpha<- overlap.tumi%>%mutate(survey="TUMI")


##### WINTER Phenology Model Runs #####
climate_dat <-readRDS(here('data/physical/climate_dat_upwelling.rds'))
season <- "Winter"
eras <- data.frame(era.region2=seq(1,9), era.region=seq(4,12))
data<- climate_dat%>%filter(season=="Winter"&region!='GoA'&Year_lag<2023)%>%
  #  filter(Year_lag!=2022&Year_lag!=2015)%>%
  distinct()%>%
  left_join(eras)

data.phe.lmW<-data%>%filter(season=="Winter")%>%
  dplyr::select(Year_lag, season, period,region,stand_tumi,stand_lusi,stand_sti,
                seasonal_NPH,seasonal_NPGO,seasonal_PDO,seasonal_ONI)%>%
  rename(NPH=seasonal_NPH,NPGO=seasonal_NPGO,PDO=seasonal_PDO,ONI=seasonal_ONI)%>%
  distinct()%>%
  pivot_longer(!c(Year_lag, season, period,region,stand_tumi,stand_lusi,stand_sti), 
               names_to = "Index_Name", values_to = "Index_Value")%>%
  pivot_longer(!c(Year_lag, season,period,region, Index_Value, Index_Name), 
               names_to = "Up_Name", values_to = "Up_Value")%>%
  distinct()
ggplot(data = data.phe.lm%>%filter(region=="Northern CC"), aes(y = Up_Value, x =Index_Value,col=as.factor(period))) +
  facet_grid(Index_Name~Up_Name, scales='free') +
  geom_point(aes(col=as.factor(period))) +
  # geom_text(aes(label=Year_lag,col=as.factor(period))) +
  geom_smooth(method = "lm", se = FALSE, aes(col=as.factor(period))) +
  geom_smooth(method = "lm", se = FALSE, col='grey') +
  scale_y_continuous(name = "Index of Abundance") +
  scale_color_manual(values =  col[1:3], name="Period",labels=c('1967 - 1988', '1989 - 2012','2013 - 2022'))+
  theme_bw()+
  xlab("Climate Index Value")+
  theme(plot.title = element_text(hjust = 0.5))+
  ggtitle("Winter")
###### STI #######
STIW<-NULL
columns<-c(which(colnames(data) == "seasonal_PDO"), which(colnames(data) == "seasonal_NPGO"),
           which(colnames(data) == "seasonal_NPH"),which(colnames(data) == "seasonal_ONI"))
for(i in 1:length(columns)){
  bayeslinmod(data, columns[i],  data$stand_sti, data$era.region2)
  STIW <-rbind(STIW,scaled.anom)
}
STIW<-STIW%>%mutate(survey="STI",era.region2=period,
                  Index=ifelse(Index=="seasonal_PDO", "PDO", ifelse(Index=="seasonal_NPGO", "NPGO",
                                                                    ifelse(Index=="seasonal_NPH", "NPH","ONI"))))%>%
  dplyr::select(!period)%>%
  left_join(data%>%dplyr::select(region,era.region2,period)%>%distinct())%>%
  mutate(Season='Winter', lag=0, era.region=era.region2)%>%
  mutate(region=fct_relevel(region,c("Northern CC","Central CC","Southern CC")))


ggplot(STIW, aes(x = beta, fill = as.factor(period))) +
  theme_bw() +
  facet_wrap(region~Index, ncol = 3, scales='free') +
  geom_density(alpha = 0.7) +
  scale_fill_manual(values = c(col[1],col[2], col[3])) +
  #theme(legend.title = element_blank(), legend.position = 'top', legend.key.size = unit(3, 'mm')) +
  geom_vline(xintercept = 0, lty = 2) +
  labs(x = "Slope",
       y = "Posterior density")

postplot(STIW, STIW$beta)
postplot(STIW,STIW$alpha)


###### LUSI #######
LUSIW<-NULL
for(i in 1:length(columns)){
  bayeslinmod(data, columns[i],  data$stand_lusi, data$era.region2)
  LUSIW <-rbind(LUSIW,scaled.anom)
}
LUSIW<-LUSIW%>%mutate(survey="LUSI",era.region2=period,
                    Index=ifelse(Index=="seasonal_PDO", "PDO", ifelse(Index=="seasonal_NPGO", "NPGO",
                                                                      ifelse(Index=="seasonal_NPH", "NPH","ONI"))))%>%
  dplyr::select(!period)%>%
  left_join(data%>%dplyr::select(region,era.region2,period)%>%distinct())%>%
  mutate(Season='Winter', lag=0, era.region=era.region2)%>%
  mutate(region=fct_relevel(region,c("Northern CC","Central CC","Southern CC")))


ggplot(LUSIW, aes(x = beta, fill = as.factor(period), group=as.factor(era.region))) +
  theme_bw() +
  facet_wrap(region~Index, ncol = 4, scales='free') +
  geom_density(alpha = 0.7) +
  scale_fill_manual(values = c(col[1],col[2], col[3])) +
  #theme(legend.title = element_blank(), legend.position = 'top', legend.key.size = unit(3, 'mm')) +
  geom_vline(xintercept = 0, lty = 2) +
  labs(x = "Slope",
       y = "Posterior density")


###### TUMI #######
TUMIW<-NULL
for(i in 1:length(columns)){
  bayeslinmod(data, columns[i],  data$stand_tumi, data$era.region2)
  TUMIW <-rbind(TUMIW,scaled.anom)
}
TUMIW<-TUMIW%>%mutate(survey="TUMI",era.region2=period,
                    Index=ifelse(Index=="seasonal_PDO", "PDO", ifelse(Index=="seasonal_NPGO", "NPGO",
                                                                      ifelse(Index=="seasonal_NPH", "NPH","ONI"))))%>%
  dplyr::select(!period)%>%
  left_join(data%>%dplyr::select(region,era.region2,period)%>%distinct())%>%
  mutate(Season='Winter', lag=0, era.region=era.region2)%>%
  mutate(region=fct_relevel(region,c("Northern CC","Central CC","Southern CC")))


ggplot(TUMIW, aes(x = beta, fill = as.factor(period), group=as.factor(era.region))) +
  theme_bw() +
  facet_wrap(region~Index, ncol = 4, scales='free') +
  geom_density(alpha = 0.7) +
  scale_fill_manual(values = c(col[1],col[2], col[3])) +
  #theme(legend.title = element_blank(), legend.position = 'top', legend.key.size = unit(3, 'mm')) +
  geom_vline(xintercept = 0, lty = 2) +
  labs(x = "Slope",
       y = "Posterior density")

#####BIOLOGICAL VERSUS UPWELLING ####
###### Spring #####


upwelling_dat <-readRDS(here('data/physical/climate_dat_upwelling.rds'))%>%
  filter(season=="Spring"&region!='GoA'&Year_lag<2023)%>%

  mutate(region=ifelse(region=='Southern CC','SCC',
                ifelse(region=="Northern CC", 'NCC','CCC')))%>%
  dplyr::select(Year_lag,  period,region, season, 
                stand_tumi,stand_lusi,stand_sti)

dfa<-readRDS(here('data/physical/climate_dat_dfa.rds'))%>%
  mutate(period=ifelse(Year_lag<1989,1,ifelse(Year_lag>2012,3,2)))%>%
    mutate(region=ifelse(trend=='CALCOFI','SCC',
                ifelse(trend=="RREAS", 'CCC',0)))%>%
  dplyr::select(Year_lag, season,region, trend,  estimate)%>%
  filter(region!=0)%>%
  distinct()%>%
  filter(season=="Spring")%>%
  merge(upwelling_dat%>%filter(season=="Spring"))%>%
  distinct()

climate_dat_RREAS<-filter(dfa, trend=="RREAS", period==2|period==3)
climate_dat_CALCOFI<-filter(dfa, trend=="CALCOFI")


climate_dat_cop<-readRDS(here('data/physical/climate_dat_cop.rds'))%>%
  filter(season=='Summer')%>%
  dplyr::select(Year_lag, seasonal_copepod_northern,seasonal_copepod_southern)%>%
  pivot_longer(!Year_lag, names_to = "trend", values_to = "estimate")%>%
  mutate(region="NCC")%>%
  merge(upwelling_dat%>%filter(season=="Spring"))
 # mutate(region="NCC", trend=ifelse(trend=='seasonal_copepod_northern', 'N. Copepod', 'S. Copepod'))
climate_dat_cop_northern<-filter(climate_dat_cop, trend=="seasonal_copepod_northern")
climate_dat_cop_southern<-filter(climate_dat_cop, trend=="seasonal_copepod_southern")

bioup_NCC_northern <-NULL
columns<-c(which(colnames(climate_dat_cop_northern) == "stand_bakun_seasonally"),
           which(colnames(climate_dat_cop_northern) == "stand_lusi"),
           which(colnames(climate_dat_cop_northern) == "stand_tumi"),
           which(colnames(climate_dat_cop_northern) == "stand_sti"))
for(i in 1:length(columns)){
  bayeslinmod(climate_dat_cop_northern, columns[i],  climate_dat_cop_northern$estimate, climate_dat_cop_northern$period)
  bioup_NCC_northern <-rbind(bioup_NCC_northern,scaled.anom)
}
bioup_NCC_northern_spring<-mutate(bioup_NCC_northern, period=ifelse(period==1,2,3),
                                  Survey="N. Copepod",region="NCC",
                                  Index=ifelse(Index=="stand_bakun_seasonally", "Upwelling", 
                                          ifelse(Index=="stand_sti", "STI",
                                              ifelse(Index=="stand_tumi", "TUMI","LUSI"))))


#bioup_NCC_northern_full <-NULL
#columns<-c(which(colnames(climate_dat_cop_northern) == "stand_bakun_seasonally"))
#for(i in 1:length(columns)){
#  bayeslinmod(climate_dat_cop_northern, columns[i],  climate_dat_cop_northern$estimate, climate_dat_cop_northern$trend)
#  bioup_NCC_northern_full <-rbind(bioup_NCC_northern_full,scaled.anom)
#}
#bioup_NCC_northern_spring_full<-mutate(bioup_NCC_northern_full, period=4,
#                                       Survey="N. Copepod",region="NCC")

 
bioup_NCC_southern <-NULL
for(i in 1:length(columns)){
  bayeslinmod(climate_dat_cop_southern, columns[i],  climate_dat_cop_southern$estimate, climate_dat_cop_southern$period)
  bioup_NCC_southern <-rbind(bioup_NCC_southern,scaled.anom)
}
bioup_NCC_southern_spring<-mutate(bioup_NCC_southern, period=ifelse(period==1,2,3),
                                  Survey="S. Copepod",region="NCC",
                                  Index=ifelse(Index=="stand_bakun_seasonally", "Upwelling", 
                                        ifelse(Index=="stand_sti", "STI",
                                        ifelse(Index=="stand_tumi", "TUMI","LUSI"))))

#bioup_NCC_southern_full <-NULL
#columns<-c(which(colnames(climate_dat_cop_southern) == "stand_bakun_seasonally"))
#for(i in 1:length(columns)){
#  bayeslinmod(climate_dat_cop_southern, columns[i],  climate_dat_cop_southern$estimate, climate_dat_cop_southern$trend)
#  bioup_NCC_southern_full <-rbind(bioup_NCC_southern_full,scaled.anom)
#}
#bioup_NCC_southern_spring_full<-mutate(bioup_NCC_southern_full, period=4,
#                           Survey="S. Copepod",region="NCC")
 
bioup_SCC <-NULL
for(i in 1:length(columns)){
  bayeslinmod(climate_dat_CALCOFI, columns[i],  climate_dat_CALCOFI$estimate, climate_dat_CALCOFI$period)
   bioup_SCC <-rbind( bioup_SCC,scaled.anom)
}
bioup_SCC_spring<-mutate(bioup_SCC,
                           Survey="CALCOFI",region="SCC",
                         Index=ifelse(Index=="stand_bakun_seasonally", "Upwelling", 
                                      ifelse(Index=="stand_sti", "STI",
                                             ifelse(Index=="stand_tumi", "TUMI","LUSI"))))

#bioup_SCC <-NULL
#columns<-c(which(colnames(climate_dat_CALCOFI) == "stand_bakun_seasonally"))
#for(i in 1:length(columns)){
#  bayeslinmod(climate_dat_CALCOFI, columns[i],  climate_dat_CALCOFI$estimate, climate_dat_CALCOFI$trend)
#   bioup_SCC <-rbind( bioup_SCC,scaled.anom)
#}
#bioup_SCC_spring_full<-mutate(bioup_SCC,Survey="CALCOFI",region="SCC", period=4)

bioup_CCC <-NULL
for(i in 1:length(columns)){
  bayeslinmod(climate_dat_RREAS, columns[i],  climate_dat_RREAS$estimate, climate_dat_RREAS$period)
  bioup_CCC <-rbind(bioup_CCC,scaled.anom)
}

bioup_CCC_spring<-mutate(bioup_CCC, period=ifelse(period==1,2,3),Survey="RREAS",region="CCC",
                         Index=ifelse(Index=="stand_bakun_seasonally", "Upwelling", 
                                      ifelse(Index=="stand_sti", "STI",
                                             ifelse(Index=="stand_tumi", "TUMI","LUSI"))))

#bioup_CCC_full <-NULL
#columns<-c(which(colnames(climate_dat_RREAS) == "stand_bakun_seasonally"))
#for(i in 1:length(columns)){
#  bayeslinmod(climate_dat_RREAS, columns[i],  climate_dat_RREAS$estimate, climate_dat_RREAS$trend)
#  bioup_CCC_full <-rbind(bioup_CCC_full,scaled.anom)
#}

#bioup_CCC_spring_full<-mutate(bioup_CCC_full, period=4,Survey="RREAS",region="CCC")


bioup_spring<- bind_rows(bioup_CCC_spring,bioup_SCC_spring,
                         bioup_NCC_southern_spring,
                         bioup_NCC_northern_spring)%>%
  mutate(season="Spring", lag=0)%>%
  mutate(Survey= ifelse(Survey=='RREAS', "RREAS (CCC)",
                    ifelse(Survey=="CALCOFI", "CALCOFI (SCC)",
                        ifelse(Survey=="N. Copepod", "N. Copepod (NCC)",
                            ifelse(Survey=="S. Copepod","S. Copepod (NCC)",Survey)))))%>%
  distinct()%>%
  mutate(Survey= fct_relevel(Survey, "N. Copepod (NCC)", 
                             "S. Copepod (NCC)","RREAS (CCC)","CALCOFI (SCC)"))


overlap.bioup <- NA
survey.names=unique(bioup_spring$Survey)
index.names=unique(bioup_spring$Index)
for(j in 1:length(survey.names)){
  temp1 <- bioup_spring%>%filter(Survey==survey.names[j])
  for(i in 1:3){
    temp <- temp1%>%filter(Index==index.names[i])%>%dplyr::select(beta, period,Survey)
   # ov1 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(alpha),temp%>%filter(period==2)%>%dplyr::select(alpha)), period1=c(1), period2=c(2), Index=index.names[i], region=region.names[j])
    #ov2 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(alpha),temp%>%filter(period==3)%>%dplyr::select(alpha)), period1=c(1), period2=c(3), Index=index.names[i], region=region.names[j])
    ov3 <- data.frame(ov=overlap(temp%>%filter(period==3)%>%dplyr::select(beta),temp%>%filter(period==2)%>%dplyr::select(beta)), period1=c(3), period2=c(2), Index=index.names[i], Survey=survey.names[j])
    #temp2<-rbind(ov1,ov2,ov3)
    overlap.bioup <-rbind(ov3,overlap.bioup)
  }
  
}


overlap.bioup.alpha <- NA
survey.names=unique(bioup_spring$Survey)
index.names=unique(bioup_spring$Index)
for(j in 1:length(survey.names)){
  temp1 <- bioup_spring%>%filter(Survey==survey.names[j])
  for(i in 1:3){
    temp <- temp1%>%filter(Index==index.names[i])%>%dplyr::select(alpha, period,Survey)
    # ov1 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(alpha),temp%>%filter(period==2)%>%dplyr::select(alpha)), period1=c(1), period2=c(2), Index=index.names[i], region=region.names[j])
    #ov2 <- data.frame(ov=overlap(temp%>%filter(period==1)%>%dplyr::select(alpha),temp%>%filter(period==3)%>%dplyr::select(alpha)), period1=c(1), period2=c(3), Index=index.names[i], region=region.names[j])
    ov3 <- data.frame(ov=overlap(temp%>%filter(period==3)%>%dplyr::select(alpha),temp%>%filter(period==2)%>%dplyr::select(alpha)), period1=c(3), period2=c(2), Index=index.names[i], Survey=survey.names[j])
    #temp2<-rbind(ov1,ov2,ov3)
    overlap.bioup.alpha <-rbind(ov3,overlap.bioup.alpha)
  }
  
}

overlap.bioup%>%filter(ov<0.1)
BioUpbeta<- ggplot(bioup_spring%>%filter(Index!="Upwelling"), aes(x = beta, fill = as.factor(period), group=as.factor(period))) +
    theme_bw() +
    facet_grid(Index~Survey, scales='free') +
    geom_density(alpha = 0.7) +
    scale_fill_manual(values = c(col[1],col[2], col[3], 'grey'),name="Period",labels=c('1967 - 1988', '1989 - 2012','2013 - 2022')) +
    #theme(legend.title = element_blank(), legend.position = 'top', legend.key.size = unit(3, 'mm')) +
    geom_vline(xintercept = 0, lty = 2) +
    labs(x = "Slope",
         y = "Posterior density")
 
 pdf(file = "Output/FigureS20_BioUpbeta.pdf",   # The directory you want to save the file in
     width = 7, # The width of the plot in inches
     height = 5)
 BioUpbeta
 dev.off()
 
 BioUpalpha<-ggplot(bioup_spring%>%filter(Index!="Upwelling"), aes(x = alpha, fill = as.factor(period), group=as.factor(period))) +
   theme_bw() +
   facet_grid(Index~Survey, scales='free') +
   geom_density(alpha = 0.7) +
   scale_fill_manual(values = c(col[1],col[2], col[3], 'grey'),name="Period",labels=c('1967 - 1988', '1989 - 2012','2013 - 2022')) +
   #theme(legend.title = element_blank(), legend.position = 'top', legend.key.size = unit(3, 'mm')) +
   geom_vline(xintercept = 0, lty = 2) +
   labs(x = "Intercept",
        y = "Posterior density")

 
 pdf(file = "Output/FigureS21_BioUpalpha.pdf",   # The directory you want to save the file in
     width = 7, # The width of the plot in inches
     height = 5)
 BioUpalpha
 dev.off()
 
 #### Full Data Results ####
 overlap.phe<-na.omit(overlap.tumi.beta)%>%
   add_row(na.omit(overlap.lusi.beta))%>%
   add_row(na.omit(overlap.sti.beta))%>%
   mutate(parameter="beta")

 overlap.phe.alpha<-na.omit(overlap.tumi.alpha)%>%
   add_row(na.omit(overlap.lusi.alpha))%>%
   add_row(na.omit(overlap.sti.alpha))%>%
   mutate(parameter="alpha")
 upwelling
 STI<-STI%>%dplyr::select(alpha,beta,Index,yfirst,ylast,survey,
                     region,period,Season,lag)%>%mutate(period=as.numeric(as.factor(period)))
 LUSI<-LUSI%>%dplyr::select(alpha,beta,Index,yfirst,ylast,survey,
                      region,period,Season,lag)%>%mutate(period=as.numeric(as.factor(period)))
 TUMI<-TUMI%>%dplyr::select(alpha,beta,Index,yfirst,ylast,survey,
                      region,period,Season,lag)%>%mutate(period=as.numeric(as.factor(period)))
 
 STIW<-STIW%>%dplyr::select(alpha,beta,Index,yfirst,ylast,survey,
                          region,period,Season,lag)%>%mutate(period=as.numeric(as.factor(period)))
 LUSIW<-LUSIW%>%dplyr::select(alpha,beta,Index,yfirst,ylast,survey,
                            region,period,Season,lag)%>%mutate(period=as.numeric(as.factor(period)))
 TUMIW<-TUMIW%>%dplyr::select(alpha,beta,Index,yfirst,ylast,survey,
                            region,period,Season,lag)%>%mutate(period=as.numeric(as.factor(period)))
 
Full_Results <- bind_rows(CALCOFI_Woffset,CALCOFI_W,CALCOFIoffset,CALCOFI,
          RREAS_Woffset,RREAS_W,RREASoffset,RREAS,
          #upwelling,
          TUMI,STI,LUSI,TUMIW,STIW,LUSIW,
          northernW%>%mutate(period=ifelse(period==1,2,ifelse(period==2,3,4))),
          northern%>%mutate(period=ifelse(period==1,2,ifelse(period==2,3,4))), 
          southernW%>%mutate(period=ifelse(period==1,2,ifelse(period==2,3,4))),
          southern%>%mutate(period=ifelse(period==1,2,ifelse(period==2,3,4))))

overlap.upW<-overlap.upW%>%mutate(region=ifelse(region=='Southern CC','SCC',
                                   ifelse(region=="Northern CC", 'NCC','CCC')),
                     season="Spring", offset=0, Survey="Upwelling")%>%
  dplyr::select(ov,period1,period2,Index,Survey,region,season,offset)
  
overlap.up<-overlap.up%>%mutate(region=ifelse(region=='Southern CC','SCC',
                                   ifelse(region=="Northern CC", 'NCC','CCC')),
                     season="Spring", offset=0, Survey="Upwelling")%>%
  dplyr::select(ov,period1,period2,Index,Survey,region,season,offset)

overlap_full.beta<-bind_rows(na.omit(overlap.CALCOFI_W.beta),
                        na.omit(overlap.CALCOFIoffset.beta),
                        na.omit(overlap.CALCOFI.beta),
                        na.omit(overlap.RREAS_Woffset.beta),
                        na.omit(overlap.RREASoffset.beta),
                        na.omit(overlap.RREAS_W.beta),
                        na.omit(overlap.RREAS.beta),
                        na.omit(overlap.northern.beta),
                        na.omit(overlap.northernW.beta),
                        na.omit(overlap.southern.beta),
                        na.omit(overlap.southernW.beta))

overlap_full.alpha<-bind_rows(na.omit(overlap.CALCOFI_W.alpha),
                             na.omit(overlap.CALCOFIoffset.alpha),
                             na.omit(overlap.CALCOFI.alpha),
                             na.omit(overlap.RREAS_Woffset.alpha),
                             na.omit(overlap.RREASoffset.alpha),
                             na.omit(overlap.RREAS_W.alpha),
                             na.omit(overlap.RREAS.alpha),
                             na.omit(overlap.northern.alpha),
                             na.omit(overlap.northernW.alpha),
                             na.omit(overlap.southern.alpha),
                             na.omit(overlap.southernW.alpha))

saveRDS(overlap_full.beta, file = 'data/overlap_Results_beta.rds')
saveRDS(overlap_full.alpha, file = 'data/overlap_Results_alpha.rds')
saveRDS(overlap.phe, file = 'data/overlap_Results_phe_beta.rds')
saveRDS(overlap.phe.alpha, file = 'data/overlap_Results_phe_alpha.rds')

saveRDS(Full_Results, file = here('data/Full_Results.rds'))

Violin_Data<-bind_rows(Full_Results,bioup_spring%>%rename(survey=Survey))%>%
mutate(survey=ifelse(survey=="N. Copepod (NCC)","N. Copepod",
                     ifelse(survey=="S. Copepod (NCC)","S. Copepod",
                            ifelse(survey=="Southern Copepod (NCC)","S. Copepod",survey))))
saveRDS(Violin_Data, file = here('data/Violin_Data.rds'))

climate_dat_cop%>%filter(region=="NCC")%>%
  dplyr::select(seasonal_ONI, Year_lag)%>%
  distinct()
  
Violin_Data%>%filter(Season=="Winter"&survey=="TUMI")

###### supplment plots spring ####
dat_plot<-climate_dat_cop%>%filter(season=="Spring")%>%
  dplyr::select(Year_lag, season, period,seasonal_copepod_northern,seasonal_copepod_southern, seasonal_NPH,seasonal_NPGO,seasonal_PDO,seasonal_ONI)%>%
  rename(NPH=seasonal_NPH,NPGO=seasonal_NPGO,PDO=seasonal_PDO,ONI=seasonal_ONI, 
          'S. Copepod (NCC)'=seasonal_copepod_southern,  'N. Copepod (NCC)'=seasonal_copepod_northern)%>%
  distinct()%>%
  pivot_longer(!c(Year_lag, season, 'N. Copepod (NCC)','S. Copepod (NCC)', period), 
               names_to = "Index_Name", values_to = "Index_Value")%>%
  pivot_longer(!c(Year_lag, season,period, Index_Value, Index_Name), 
               names_to = "trend", values_to = "estimate")%>%
  distinct()%>%
bind_rows(dat%>%filter(season=="Spring")%>%
  dplyr::select(Year_lag, season, period,estimate, trend,seasonal_NPH,seasonal_NPGO,seasonal_PDO,seasonal_ONI)%>%
  rename(NPH=seasonal_NPH,NPGO=seasonal_NPGO,PDO=seasonal_PDO,ONI=seasonal_ONI)%>%
  distinct()%>%
  pivot_longer(!c(Year_lag, season, period,estimate, trend), 
               names_to = "Index_Name", values_to = "Index_Value")%>%
    filter(trend=="RREAS"|trend=="CALCOFI")%>%
    mutate(trend= ifelse(trend=='RREAS', "RREAS (CCC)",ifelse(trend=="CALCOFI", "CALCOFI (SCC)", NA)))%>%
  distinct())%>%
  mutate(trend = fct_relevel(trend, "N. Copepod (NCC)", 
                             "S. Copepod (NCC)","RREAS (CCC)","CALCOFI (SCC)"))

dat_plot
bio_dat<-ggplot(data = dat_plot%>%filter(season=="Spring"&period!=4)%>%filter(!c(trend=="RREAS (CCC)"&period==1)), 
       aes(y = estimate, x =Index_Value,col=as.factor(period))) +
  facet_grid(Index_Name~trend, scales='free') +
  geom_point(aes(col=as.factor(period)), alpha = 0.5) +
  # geom_text(aes(label=Year_lag,col=as.factor(period))) +
  geom_smooth(method = "lm", se = FALSE, aes(col=as.factor(period))) +
  #geom_smooth(method = "lm", se = FALSE, col='grey') +
  scale_y_continuous(name = "Index of Abundance") +
  scale_color_manual(values =  col[1:3], name="Period",labels=c('1967 - 1988', '1989 - 2012','2013 - 2022'))+
  theme_bw()+
  xlab("Climate Index Value")+
  theme(plot.title = element_text(hjust = 0.5))+
  ggtitle("Spring")
bio_dat
pdf(file = "Output/FigureS16_BIOlinearregression.pdf",   # The directory you want to save the file in
    width = 7, # The width of the plot in inches
    height = 5)
bio_dat
#dev.off()


bio<-Full_Results%>%
mutate(survey= ifelse(survey=='RREAS', "RREAS (CCC)",
                      ifelse(survey=="CALCOFI", "CALCOFI (SCC)",
                        ifelse(survey=="N. Copepod", "N. Copepod (NCC)",
                               ifelse(survey=="Southern Copepod (NCC)","S. Copepod (NCC)",survey)))))%>%
  distinct()%>%
  mutate(survey = fct_relevel(survey, "N. Copepod (NCC)", 
                             "S. Copepod (NCC)","RREAS (CCC)","CALCOFI (SCC)"))
bio.beta<-ggplot(data = bio%>%
                  filter(Season=="Spring"&period!=4&lag==0)%>%
         filter(survey=="RREAS (CCC)"|survey=="CALCOFI (SCC)"|survey=="N. Copepod (NCC)"|survey=="S. Copepod (NCC)"),
       aes(x = beta, fill = as.factor(period))) +
  theme_bw() +
  facet_grid(Index~survey, scales='free') +
  geom_density(alpha = 0.7) +
#  xlim(c(-1,1))+
  scale_fill_manual(values = c(col[1],col[2], col[3]), name="Period",labels=c('1967 - 1988', '1989 - 2012','2013 - 2022')) +
  #theme(legend.title = element_blank(), legend.position = 'top', legend.key.size = unit(3, 'mm')) +
  geom_vline(xintercept = 0, lty = 2) +
  labs(x = "Slope",
       y = "Posterior density")
pdf(file = "Output/FigureS17_Biobeta.pdf",   # The directory you want to save the file in
    width = 7, # The width of the plot in inches
    height = 5)
bio.beta
dev.off()

bio.alpha<-ggplot(data = bio%>%
                   filter(Season=="Spring"&period!=4&lag==0)%>%
                   filter(survey=="RREAS (CCC)"|survey=="CALCOFI (SCC)"|survey=="N. Copepod (NCC)"|survey=="S. Copepod (NCC)"),
                 aes(x = alpha, fill = as.factor(period))) +
  theme_bw() +
  facet_grid(Index~survey, scales='free') +
  geom_density(alpha = 0.7) +
  #  xlim(c(-1,1))+
  scale_fill_manual(values = c(col[1],col[2], col[3]), name="Period",labels=c('1967 - 1988', '1989 - 2012','2013 - 2022')) +
  #theme(legend.title = element_blank(), legend.position = 'top', legend.key.size = unit(3, 'mm')) +
  geom_vline(xintercept = 0, lty = 2) +
  labs(x = "Intercept",
       y = "Posterior density")
pdf(file = "Output/FigureS18_Bioalpha.pdf",   # The directory you want to save the file in
    width = 7, # The width of the plot in inches
    height = 5)
bio.alpha
dev.off()


###### supplment plots winter ####
dat_plot<-climate_dat_cop%>%filter(season=="Winter")%>%
  dplyr::select(Year_lag, season, period,seasonal_copepod_northern,seasonal_copepod_southern, seasonal_NPH,seasonal_NPGO,seasonal_PDO,seasonal_ONI)%>%
  rename(NPH=seasonal_NPH,NPGO=seasonal_NPGO,PDO=seasonal_PDO,ONI=seasonal_ONI, 
         'S. Copepod (NCC)'=seasonal_copepod_southern,  'N. Copepod (NCC)'=seasonal_copepod_northern)%>%
  distinct()%>%
  pivot_longer(!c(Year_lag, season, 'N. Copepod (NCC)','S. Copepod (NCC)', period), 
               names_to = "Index_Name", values_to = "Index_Value")%>%
  pivot_longer(!c(Year_lag, season,period, Index_Value, Index_Name), 
               names_to = "trend", values_to = "estimate")%>%
  distinct()%>%
  bind_rows(dat%>%filter(season=="Winter")%>%
              dplyr::select(Year_lag, season, period,estimate, trend,seasonal_NPH,seasonal_NPGO,seasonal_PDO,seasonal_ONI)%>%
              rename(NPH=seasonal_NPH,NPGO=seasonal_NPGO,PDO=seasonal_PDO,ONI=seasonal_ONI)%>%
              distinct()%>%
              pivot_longer(!c(Year_lag, season, period,estimate, trend), 
                           names_to = "Index_Name", values_to = "Index_Value")%>%
              filter(trend=="RREAS"|trend=="CALCOFI")%>%
              mutate(trend= ifelse(trend=='RREAS', "RREAS (CCC)",ifelse(trend=="CALCOFI", "CALCOFI (SCC)", NA)))%>%
              distinct())%>%
  mutate(trend = fct_relevel(trend, "N. Copepod (NCC)", 
                             "S. Copepod (NCC)","RREAS (CCC)","CALCOFI (SCC)"))


bio_dat<-ggplot(data = dat_plot%>%filter(season=="Winter"&period!=4), 
                aes(y = estimate, x =Index_Value,col=as.factor(period))) +
  facet_grid(Index_Name~trend, scales='free') +
  geom_point(aes(col=as.factor(period)), alpha = 0.5) +
  # geom_text(aes(label=Year_lag,col=as.factor(period))) +
  geom_smooth(method = "lm", se = FALSE, aes(col=as.factor(period))) +
  #geom_smooth(method = "lm", se = FALSE, col='grey') +
  scale_y_continuous(name = "Index of Abundance") +
  scale_color_manual(values =  col[1:3], name="Period",labels=c('1967 - 1988', '1989 - 2012','2013 - 2022'))+
  theme_bw()+
  xlab("Climate Index Value")+
  theme(plot.title = element_text(hjust = 0.5))+
  ggtitle("Winter")
bio_dat
pdf(file = "Output/FigureS31_BIOlinearregressionW.pdf",   # The directory you want to save the file in
    width =7, # The width of the plot in inches
    height = 5)
bio_dat
dev.off()


bio<-Full_Results%>%
  mutate(survey= ifelse(survey=='RREAS', "RREAS (CCC)",
                        ifelse(survey=="CALCOFI", "CALCOFI (SCC)",
                               ifelse(survey=="N. Copepod", "N. Copepod (NCC)",
                                      ifelse(survey=="Southern Copepod (NCC)","S. Copepod (NCC)",survey)))))%>%
  distinct()%>%
  mutate(survey = fct_relevel(survey, "N. Copepod (NCC)", 
                              "S. Copepod (NCC)","RREAS (CCC)","CALCOFI (SCC)"))
bio.beta<-ggplot(data = bio%>%
                   filter(Season=="Winter"&period!=4&lag==0)%>%
                   filter(survey=="RREAS (CCC)"|survey=="CALCOFI (SCC)"|survey=="N. Copepod (NCC)"|survey=="S. Copepod (NCC)"),
                 aes(x = beta, fill = as.factor(period))) +
  theme_bw() +
  facet_grid(Index~survey, scales='free') +
  geom_density(alpha = 0.7) +
  #  xlim(c(-1,1))+
  scale_fill_manual(values = c(col[1],col[2], col[3]), name="Period",labels=c('1967 - 1988', '1989 - 2012','2013 - 2022')) +
  #theme(legend.title = element_blank(), legend.position = 'top', legend.key.size = unit(3, 'mm')) +
  geom_vline(xintercept = 0, lty = 2) +
  labs(x = "Slope",
       y = "Posterior density")
pdf(file = "Output/FigureS32_BiobetaW.pdf",   # The directory you want to save the file in
    width = 7, # The width of the plot in inches
    height = 5)
bio.beta
dev.off()

bio.alpha<-ggplot(data = bio%>%
                    filter(Season=="Winter"&period!=4&lag==0)%>%
                    filter(survey=="RREAS (CCC)"|survey=="CALCOFI (SCC)"|survey=="N. Copepod (NCC)"|survey=="S. Copepod (NCC)"),
                  aes(x = alpha, fill = as.factor(period))) +
  theme_bw() +
  facet_grid(Index~survey, scales='free') +
  geom_density(alpha = 0.7) +
  #  xlim(c(-1,1))+
  scale_fill_manual(values = c(col[1],col[2], col[3]), name="Period",labels=c('1967 - 1988', '1989 - 2012','2013 - 2022')) +
  #theme(legend.title = element_blank(), legend.position = 'top', legend.key.size = unit(3, 'mm')) +
  geom_vline(xintercept = 0, lty = 2) +
  labs(x = "Intercept",
       y = "Posterior density")
pdf(file = "Output/FigureS33_BioalphaW.pdf",   # The directory you want to save the file in
    width = 7, # The width of the plot in inches
    height = 5)
bio.alpha
dev.off()

data<- climate_dat%>%filter(season=="Winter"&region!='GoA'&Year_lag<2023&period!=4)%>%
  #  filter(Year_lag!=2022&Year_lag!=2015)%>%
  distinct()%>%
  left_join(eras)%>%
  mutate(region=fct_relevel(region,c("Northern CC","Central CC","Southern CC")))
index.names <-c("PDO", "NPGO", "NPH", "ONI")
region.names<-unique(climate_dat$region)
period.names <- unique(climate_dat$period)

data.phe.lm<-data%>%filter(season=="Winter")%>%
  dplyr::select(Year_lag, season, period,region,stand_tumi,stand_lusi,stand_sti,stand_bakun_seasonally,
                seasonal_NPH,seasonal_NPGO,seasonal_PDO,seasonal_ONI)%>%
  rename(NPH=seasonal_NPH,NPGO=seasonal_NPGO,PDO=seasonal_PDO,ONI=seasonal_ONI)%>%
  distinct()%>%
  pivot_longer(!c(Year_lag, season, period,region,stand_tumi,stand_lusi,stand_sti,stand_bakun_seasonally), 
               names_to = "Index_Name", values_to = "Index_Value")%>%
  pivot_longer(!c(Year_lag, season,period,region, Index_Value, Index_Name), 
               names_to = "Up_Name", values_to = "Up_Value")%>%
  distinct()

sti_datW <-ggplot(data = data.phe.lm%>%filter(Up_Name=="stand_sti"&period!=4), aes(y = Up_Value, x =Index_Value,col=as.factor(period))) +
  facet_grid(Index_Name~region, scales='free') +
  geom_point(aes(col=as.factor(period)), alpha = 0.5) +
  # geom_text(aes(label=Year_lag,col=as.factor(period))) +
  geom_smooth(method = "lm", se = FALSE, aes(col=as.factor(period))) +
  #geom_smooth(method = "lm", se = FALSE, col='grey') +
  scale_y_continuous(name = "STI") +
  scale_color_manual(values =  col[1:3], name="Period",labels=c('1967 - 1988', '1989 - 2012','2013 - 2022'))+
  theme_bw()+
  xlab("Climate Index Value")+
  theme(plot.title = element_text(hjust = 0.5))+
  ggtitle("Winter")
sti_datW
pdf(file = "Output/FigureS25_STIlinearregressionW.pdf",   # The directory you want to save the file in
    width = 7, # The width of the plot in inches
    height = 5)
sti_datW
dev.off()


tumi_datW <-ggplot(data = data.phe.lm%>%filter(Up_Name=="stand_tumi"&period!=4), aes(y = Up_Value, x =Index_Value,col=as.factor(period))) +
  facet_grid(Index_Name~region, scales='free') +
  geom_point(aes(col=as.factor(period)), alpha = 0.5) +
  # geom_text(aes(label=Year_lag,col=as.factor(period))) +
  geom_smooth(method = "lm", se = FALSE, aes(col=as.factor(period))) +
  #geom_smooth(method = "lm", se = FALSE, col='grey') +
  scale_y_continuous(name = "TUMI") +
  scale_color_manual(values =  col[1:3], name="Period",labels=c('1967 - 1988', '1989 - 2012','2013 - 2022'))+
  theme_bw()+
  xlab("Climate Index Value")+
  theme(plot.title = element_text(hjust = 0.5))+
  ggtitle("Winter")
tumi_datW
pdf(file = "Output/FigureS22_TUMIlinearregressionW.pdf",   # The directory you want to save the file in
    width = 7, # The width of the plot in inches
    height = 5)
tumi_datW
dev.off()


lusi_datW <-ggplot(data = data.phe.lm%>%filter(Up_Name=="stand_lusi"&period!=4), aes(y = Up_Value, x =Index_Value,col=as.factor(period))) +
  facet_grid(Index_Name~region, scales='free') +
  geom_point(aes(col=as.factor(period)), alpha = 0.5) +
  # geom_text(aes(label=Year_lag,col=as.factor(period))) +
  geom_smooth(method = "lm", se = FALSE, aes(col=as.factor(period))) +
  #geom_smooth(method = "lm", se = FALSE, col='grey') +
  scale_y_continuous(name = "LUSI") +
  scale_color_manual(values =  col[1:3], name="Period",labels=c('1967 - 1988', '1989 - 2012','2013 - 2022'))+
  theme_bw()+
  xlab("Climate Index Value")+
  theme(plot.title = element_text(hjust = 0.5))+
  ggtitle("Winter")
lusi_datW
pdf(file = "Output/FigureS28_lusilinearregressionW.pdf",   # The directory you want to save the file in
    width = 7, # The width of the plot in inches
    height = 5)
lusi_datW
dev.off()

STI_betaW <-ggplot(STIW, aes(x = beta, fill = as.factor(period))) +
  theme_bw() +
  facet_grid(Index~region) +
  geom_density(alpha = 0.7) +
  scale_fill_manual(values = c(col[1],col[2], col[3]), name="Period",labels=c('1967 - 1988', '1989 - 2012','2013 - 2022')) +
  #theme(legend.title = element_blank(), legend.position = 'top', legend.key.size = unit(3, 'mm')) +
  geom_vline(xintercept = 0, lty = 2) +
  labs(x = "Slope",
       y = "Posterior density")

pdf(file = "Output/FigureS26_STIbetaW.pdf",   # The directory you want to save the file in
    width = 7, # The width of the plot in inches
    height = 5)
STI_betaW
dev.off()

STI_alphaW <-ggplot(STIW, aes(x = alpha, fill = as.factor(period))) +
  theme_bw() +
  facet_grid(Index~region) +
  geom_density(alpha = 0.7) +
  scale_fill_manual(values = c(col[1],col[2], col[3]), name="Period",labels=c('1967 - 1988', '1989 - 2012','2013 - 2022')) +
  #theme(legend.title = element_blank(), legend.position = 'top', legend.key.size = unit(3, 'mm')) +
  geom_vline(xintercept = 0, lty = 2) +
  labs(x = "Intercept",
       y = "Posterior density")

pdf(file = "Output/FigureS27_STIalphaW.pdf",   # The directory you want to save the file in
    width = 7, # The width of the plot in inches
    height = 5)
STI_alphaW
dev.off()

TUMI_betaW <-ggplot(TUMIW, aes(x = beta, fill = as.factor(period))) +
  theme_bw() +
  facet_grid(Index~region) +
  geom_density(alpha = 0.7) +
  scale_fill_manual(values = c(col[1],col[2], col[3]), name="Period",labels=c('1967 - 1988', '1989 - 2012','2013 - 2022')) +
  #theme(legend.title = element_blank(), legend.position = 'top', legend.key.size = unit(3, 'mm')) +
  geom_vline(xintercept = 0, lty = 2) +
  labs(x = "Slope",
       y = "Posterior density")

pdf(file = "Output/FigureS23_TUMIbetaW.pdf",   # The directory you want to save the file in
    width = 7, # The width of the plot in inches
    height = 5)
TUMI_betaW
dev.off()

TUMI_alphaW <-ggplot(TUMIW, aes(x = alpha, fill = as.factor(period))) +
  theme_bw() +
  facet_grid(Index~region) +
  geom_density(alpha = 0.7) +
  scale_fill_manual(values = c(col[1],col[2], col[3]), name="Period",labels=c('1967 - 1988', '1989 - 2012','2013 - 2022')) +
  #theme(legend.title = element_blank(), legend.position = 'top', legend.key.size = unit(3, 'mm')) +
  geom_vline(xintercept = 0, lty = 2) +
  labs(x = "Intercept",
       y = "Posterior density")

pdf(file = "Output/FigureS24_TUMIalphaW.pdf",   # The directory you want to save the file in
    width = 7, # The width of the plot in inches
    height = 5)
TUMI_alphaW
dev.off()


LUSI_betaW <-ggplot(LUSIW, aes(x = beta, fill = as.factor(period))) +
  theme_bw() +
  facet_grid(Index~region) +
  geom_density(alpha = 0.7) +
  scale_fill_manual(values = c(col[1],col[2], col[3]), name="Period",labels=c('1967 - 1988', '1989 - 2012','2013 - 2022')) +
  #theme(legend.title = element_blank(), legend.position = 'top', legend.key.size = unit(3, 'mm')) +
  geom_vline(xintercept = 0, lty = 2) +
  labs(x = "Slope",
       y = "Posterior density")

pdf(file = "Output/FigureS29_LUSIbetaW.pdf",   # The directory you want to save the file in
    width = 7, # The width of the plot in inches
    height = 5)
LUSI_betaW
dev.off()

LUSI_alphaW <-ggplot(LUSIW, aes(x = alpha, fill = as.factor(period))) +
  theme_bw() +
  facet_grid(Index~region) +
  geom_density(alpha = 0.7) +
  scale_fill_manual(values = c(col[1],col[2], col[3]), name="Period",labels=c('1967 - 1988', '1989 - 2012','2013 - 2022')) +
  #theme(legend.title = element_blank(), legend.position = 'top', legend.key.size = unit(3, 'mm')) +
  geom_vline(xintercept = 0, lty = 2) +
  labs(x = "Intercept",
       y = "Posterior density")

pdf(file = "Output/FigureS30_LUSIalphaW.pdf",   # The directory you want to save the file in
    width = 7, # The width of the plot in inches
    height = 5)
LUSI_alphaW
dev.off()

#### Temporal Lag ####
dat_plot<-dat%>%filter(season=="Spring"&trend!="SEA")%>%
  dplyr::select(Year_lag, estimateoffset1,season, period,trend, seasonal_NPH,seasonal_NPGO,seasonal_PDO,seasonal_ONI)%>%
  rename(NPH=seasonal_NPH,NPGO=seasonal_NPGO,PDO=seasonal_PDO,ONI=seasonal_ONI)%>%
  distinct()%>%
  pivot_longer(!c(Year_lag,estimateoffset1,trend, season, period), 
               names_to = "Index_Name", values_to = "Index_Value")%>%
  distinct()%>%
              mutate(trend= ifelse(trend=='RREAS', "RREAS (CCC)",ifelse(trend=="CALCOFI", "CALCOFI (SCC)", NA)))%>%
              distinct()

bio_dat<-ggplot(data =na.omit(dat_plot%>%filter(season=="Spring")), 
                aes(y = estimateoffset1, x =Index_Value,col=as.factor(period))) +
  facet_grid(Index_Name~trend, scales='free') +
  geom_point(aes(col=as.factor(period)), alpha = 0.5) +
  # geom_text(aes(label=Year_lag,col=as.factor(period))) +
  geom_smooth(method = "lm", se = FALSE, aes(col=as.factor(period))) +
  #geom_smooth(method = "lm", se = FALSE, col='grey') +
  scale_y_continuous(name = "Index of Abundance") +
  scale_color_manual(values =  col[1:3], name="Period",labels=c('1967 - 1988', '1989 - 2012','2013 - 2022'))+
  theme_bw()+
  xlab("Climate Index Value")+
  theme(plot.title = element_text(hjust = 0.5))+
  ggtitle("1-year Lag")
bio_dat
pdf(file = "Output/FigureS33_BIOlinearregressionLAG.pdf",   # The directory you want to save the file in
    width = 7, # The width of the plot in inches
    height = 5)
bio_dat
dev.off()


bio<-Full_Results%>%
  mutate(survey= ifelse(survey=='RREAS', "RREAS (CCC)",
                        ifelse(survey=="CALCOFI", "CALCOFI (SCC)",
                               ifelse(survey=="N. Copepod", "N. Copepod (NCC)",
                                      ifelse(survey=="Southern Copepod (NCC)","S. Copepod (NCC)",survey)))))%>%
  distinct()%>%
  mutate(survey = fct_relevel(survey, "N. Copepod (NCC)", 
                              "S. Copepod (NCC)","RREAS (CCC)","CALCOFI (SCC)"))
bio.beta<-ggplot(data = bio%>%
                   filter(Season=="Spring"&period!=4&lag==1)%>%
                   filter(survey=="RREAS (CCC)"|survey=="CALCOFI (SCC)"|survey=="N. Copepod (NCC)"|survey=="S. Copepod (NCC)"),
                 aes(x = beta, fill = as.factor(period))) +
  theme_bw() +
  facet_grid(Index~survey, scales='free') +
  geom_density(alpha = 0.7) +
  #  xlim(c(-1,1))+
  scale_fill_manual(values = c(col[1],col[2], col[3]), name="Period",labels=c('1967 - 1988', '1989 - 2012','2013 - 2022')) +
  #theme(legend.title = element_blank(), legend.position = 'top', legend.key.size = unit(3, 'mm')) +
  geom_vline(xintercept = 0, lty = 2) +
  labs(x = "Slope",
       y = "Posterior density")
pdf(file = "Output/FigureS34_Biobeta.pdf",   # The directory you want to save the file in
    width = 7, # The width of the plot in inches
    height = 5)
bio.beta
dev.off()

bio.alpha<-ggplot(data = bio%>%
                    filter(Season=="Spring"&period!=4&lag==1)%>%
                    filter(survey=="RREAS (CCC)"|survey=="CALCOFI (SCC)"|survey=="N. Copepod (NCC)"|survey=="S. Copepod (NCC)"),
                  aes(x = alpha, fill = as.factor(period))) +
  theme_bw() +
  facet_grid(Index~survey, scales='free') +
  geom_density(alpha = 0.7) +
  #  xlim(c(-1,1))+
  scale_fill_manual(values = c(col[1],col[2], col[3]), name="Period",labels=c('1967 - 1988', '1989 - 2012','2013 - 2022')) +
  #theme(legend.title = element_blank(), legend.position = 'top', legend.key.size = unit(3, 'mm')) +
  geom_vline(xintercept = 0, lty = 2) +
  labs(x = "Intercept",
       y = "Posterior density")
pdf(file = "Output/FigureS34_Bioalpha.pdf",   # The directory you want to save the file in
    width = 7, # The width of the plot in inches
    height = 5)
bio.alpha
dev.off()


dat_plot2<-climate_dat_cop%>%
  bind_rows(climate_dat_CALCOFI)%>%
  bind_rows(climate_dat_RREAS)%>%
 # bind_rows(filter(dfa, trend=="RREAS"))%>%
  rename(TUMI=stand_tumi, STI=stand_sti, LUSI=stand_lusi)%>%
  mutate(trend=ifelse(trend=='seasonal_copepod_northern', 'N. Copepod (NCC)',
                      ifelse(trend=='seasonal_copepod_southern', 'S. Copepod (NCC)',
                      ifelse(trend=="RREAS", "RREAS (CCC)", "CALCOFI (SCC)"))))%>%
  pivot_longer(c(TUMI, LUSI,STI),names_to = 'Index_Name', values_to = 'Index_Value')%>%
  mutate(trend = fct_relevel(trend, "N. Copepod (NCC)", 
                              "S. Copepod (NCC)","RREAS (CCC)","CALCOFI (SCC)"))

up_bio<-ggplot(data = dat_plot2%>%filter(period!=4), 
                aes(y = estimate, x =Index_Value,col=as.factor(period))) +
  facet_grid(trend~Index_Name, scales='free') +
  geom_point(aes(col=as.factor(period)), alpha = 0.5) +
  # geom_text(aes(label=Year_lag,col=as.factor(period))) +
  geom_smooth(method = "lm", se = FALSE, aes(col=as.factor(period))) +
  #geom_smooth(method = "lm", se = FALSE, col='grey') +
  scale_y_continuous(name = "Index of Abundance") +
  scale_color_manual(values =  col[1:3], name="Period",labels=c('1967 - 1988', '1989 - 2012','2013 - 2022'))+
  theme_bw()+
  xlab("Climate Index Value")+
  theme(plot.title = element_text(hjust = 0.5))
pdf(file = "Output/FigureS10_upwellingbio.pdf",   # The directory you want to save the file in
    width = 7, # The width of the plot in inches
    height = 5)
up_bio
dev.off()
